refactor: load Valkey function library from glidemq.lua
Keep the Lua source in a real file so later coverage can measure FCALL paths, and copy it into dist on build. FUNCTION LOAD bytes stay the same aside from the version placeholder substitution.
