+++ src/functions/glidemq.lua
@@ -1,12 +1,19 @@
 #!lua name=glidemq
 
--- glide-mq function library v1
--- Loaded once via FUNCTION LOAD, persistent across restarts.
--- All keys use hash tag {queueName} for cluster safety.
+local PRIORITY_SHIFT = 4398046511104
 
-local PRIORITY_SHIFT = 4398046511104 -- 2^42
+-- Guarded decrement of the per-queue list-active counter.
+-- Used at every site that tracks completion/failure/release of a list-sourced
+-- job (entryId == ''). The guard is required because healListActive cannot
+-- recover from a negative counter (its early-out at <= 0 is intentional - it
+-- only repairs positive drift from worker crashes), so a duplicate decrement
+-- would latch the counter negative forever.
+local function decrListActive(listActiveKey)
+  if not listActiveKey or listActiveKey == '' then return end
+  local la = tonumber(redis.call('GET', listActiveKey)) or 0
+  if la > 0 then redis.call('DECR', listActiveKey) end
+end
 
--- Helper: emit an event to the events stream
 local function emitEvent(eventsKey, eventType, jobId, extraFields)
   local fields = {'event', eventType, 'jobId', tostring(jobId)}
   if extraFields then
@@ -18,12 +25,20 @@ local function emitEvent(eventsKey, eventType, jobId, extraFields)
   redis.call('XADD', eventsKey, 'MAXLEN', '~', '1000', '*', unpack(fields))
 end
 
-local function markOrderingDone(jobKey, jobId)
-  local orderingKey = redis.call('HGET', jobKey, 'orderingKey')
+local function markOrderingDone(jobKey, jobId, hintOrderingKey, hintOrderingSeq)
+  local orderingKey = hintOrderingKey
+  if not orderingKey or orderingKey == '' then
+    orderingKey = redis.call('HGET', jobKey, 'orderingKey')
+  end
   if not orderingKey or orderingKey == '' then
     return
   end
-  local orderingSeq = tonumber(redis.call('HGET', jobKey, 'orderingSeq')) or 0
+  local orderingSeq = nil
+  if hintOrderingSeq ~= nil and hintOrderingSeq ~= '' then
+    orderingSeq = tonumber(hintOrderingSeq) or 0
+  else
+    orderingSeq = tonumber(redis.call('HGET', jobKey, 'orderingSeq')) or 0
+  end
   if orderingSeq <= 0 then
     return
   end
@@ -54,6 +69,272 @@ local function markOrderingDone(jobKey, jobId)
   end
 end
 
+local MAX_SKIP_ADVANCE_STEPS = 128
+
+-- Advance nextSeq past any skip markers left by debounce on ordered jobs.
+-- Skip markers are hash fields 'skip:<seq>' set when debounce deletes an ordered job.
+-- Returns the advanced nextSeq. Cleans up markers via HDEL as it goes.
+-- Work is bounded per call to avoid long synchronous loops.
+local function advancePastSkips(groupHashKey, nextSeq)
+  local steps = 0
+  while steps < MAX_SKIP_ADVANCE_STEPS and redis.call('HDEL', groupHashKey, 'skip:' .. tostring(nextSeq)) == 1 do
+    nextSeq = nextSeq + 1
+    steps = steps + 1
+  end
+  return nextSeq
+end
+
+-- Refill token bucket using remainder accumulator for precision.
+-- tbRefillRate is in millitokens/second. Returns current millitokens after refill.
+-- Side effect: updates tbTokens, tbLastRefill, tbRefillRemainder on the group hash.
+local function tbRefill(groupHashKey, g, now)
+  local tbCapacity = tonumber(g.tbCapacity) or 0
+  if tbCapacity <= 0 then return 0 end
+  local tbTokens = tonumber(g.tbTokens) or tbCapacity
+  if tbTokens >= tbCapacity then return tbCapacity end
+  local tbRefillRate = tonumber(g.tbRefillRate) or 0
+  local tbLastRefill = tonumber(g.tbLastRefill) or now
+  local tbRefillRemainder = tonumber(g.tbRefillRemainder) or 0
+  local elapsed = now - tbLastRefill
+  if elapsed <= 0 or tbRefillRate <= 0 then return tbTokens end
+  -- Cap elapsed to prevent overflow in long-idle buckets
+  local maxElapsed = math.ceil(tbCapacity * 1000 / tbRefillRate)
+  if elapsed > maxElapsed then elapsed = maxElapsed end
+  local raw = elapsed * tbRefillRate + tbRefillRemainder
+  local added = math.floor(raw / 1000)
+  local newRemainder = raw % 1000
+  local newTokens = math.min(tbCapacity, tbTokens + added)
+  redis.call('HSET', groupHashKey,
+    'tbTokens', tostring(newTokens),
+    'tbLastRefill', tostring(now),
+    'tbRefillRemainder', tostring(newRemainder))
+  return newTokens
+end
+
+local function recordMetrics(metricsKey, timestamp, duration)
+  local minuteTs = timestamp - (timestamp % 60000)
+  local newCount = tonumber(redis.call('HINCRBY', metricsKey, 'm:' .. minuteTs .. ':c', 1))
+  if duration > 0 then
+    redis.call('HINCRBY', metricsKey, 'm:' .. minuteTs .. ':d', duration)
+  end
+  if newCount and (newCount % 1000 == 0) then
+    local cutoff = minuteTs - 86400000
+    local fields = redis.call('HKEYS', metricsKey)
+    local toDelete = {}
+    for _, f in ipairs(fields) do
+      local ts = tonumber(string.match(f, '^m:(%d+):'))
+      if ts and ts < cutoff then
+        toDelete[#toDelete + 1] = f
+      end
+    end
+    if #toDelete > 0 and #toDelete <= 1000 then
+      redis.call('HDEL', metricsKey, unpack(toDelete))
+    elseif #toDelete > 1000 then
+      for i = 1, #toDelete, 1000 do
+        redis.call('HDEL', metricsKey, unpack(toDelete, i, math.min(i + 999, #toDelete)))
+      end
+    end
+  end
+end
+
+local function xaddJob(streamKey, jobId, jobName)
+  redis.call('XADD', streamKey, '*', 'jobId', jobId, 'name', jobName or '')
+end
+local function groupqScore(jobKey, jobId)
+  local seq = tonumber(redis.call('HGET', jobKey, 'orderingSeq'))
+  if seq and seq > 0 then return seq end
+  local num = tonumber(jobId)
+  if num then return num end
+  -- String IDs: use timestamp from job hash to preserve insertion order
+  local ts = tonumber(redis.call('HGET', jobKey, 'timestamp'))
+  if ts then return ts end
+  return 0
+end
+
+
+local function releaseGroupSlotAndPromote(jobKey, jobId, now, hintGroupKey)
+  local gk = hintGroupKey
+  if not gk or gk == '' then
+    gk = redis.call('HGET', jobKey, 'groupKey')
+  end
+  if not gk or gk == '' then return end
+  local prefix = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
+  local groupHashKey = prefix .. 'group:' .. gk
+  -- Load all group fields in one call
+  local gFields = redis.call('HGETALL', groupHashKey)
+  local g = {}
+  for gf = 1, #gFields, 2 do g[gFields[gf]] = gFields[gf + 1] end
+  local cur = tonumber(g.active) or 0
+  local newActive = (cur > 0) and (cur - 1) or 0
+  if cur > 0 then
+    redis.call('HSET', groupHashKey, 'active', tostring(newActive))
+  end
+  local waitListKey = prefix .. 'groupq:' .. gk
+  local waitLen = redis.call('ZCARD', waitListKey)
+  if waitLen == 0 then return end
+  -- Concurrency gate: if still at or above max after decrement, do not promote
+  local maxConc = tonumber(g.maxConcurrency) or 0
+  if maxConc > 0 and newActive >= maxConc then return end
+  -- Rate limit gate (skip if now is nil or 0 for safe fallback)
+  -- Only blocks promotion; does NOT increment rateCount. moveToActive handles counting.
+  local rateMax = tonumber(g.rateMax) or 0
+  local rateRemaining = 0
+  local ts = tonumber(now) or 0
+  if ts > 0 and rateMax > 0 then
+    local rateDuration = tonumber(g.rateDuration) or 0
+    if rateDuration > 0 then
+      local rateWindowStart = tonumber(g.rateWindowStart) or 0
+      local rateCount = tonumber(g.rateCount) or 0
+      if ts - rateWindowStart < rateDuration then
+        if rateCount >= rateMax then
+          -- Window active and at capacity: do not promote, register for scheduler
+          local rateLimitedKey = prefix .. 'ratelimited'
+          redis.call('ZADD', rateLimitedKey, rateWindowStart + rateDuration, gk)
+          return
+        end
+        rateRemaining = rateMax - rateCount
+      end
+    end
+  end
+  -- Token bucket gate: check head job cost before promoting
+  local tbCap = tonumber(g.tbCapacity) or 0
+  if ts > 0 and tbCap > 0 then
+    local tbTokensCur = tbRefill(groupHashKey, g, ts)
+    -- Peek at head job, skipping tombstones and DLQ'd jobs (up to 10 iterations)
+    local tbCheckPasses = 0
+    local tbOk = false
+    while tbCheckPasses < 10 do
+      tbCheckPasses = tbCheckPasses + 1
+      local headMembers = redis.call('ZRANGE', waitListKey, 0, 0)
+      local headJobId = headMembers[1]
+      if not headJobId then break end
+      local headJobKey = prefix .. 'job:' .. headJobId
+      -- Tombstone guard: job hash deleted - remove and check next
+      if redis.call('EXISTS', headJobKey) == 0 then
+        redis.call('ZREM', waitListKey, headJobId)
+      else
+        local headCost = tonumber(redis.call('HGET', headJobKey, 'cost')) or 1000
+        -- DLQ guard: cost > capacity - remove, fail, check next
+        if headCost > tbCap then
+          local metricsKey = prefix .. 'metrics:failed'
+          local processedOn = tonumber(redis.call('HGET', headJobKey, 'processedOn')) or ts
+          redis.call('ZREM', waitListKey, headJobId)
+          redis.call('ZADD', prefix .. 'failed', ts, headJobId)
+          redis.call('HSET', headJobKey,
+            'state', 'failed',
+            'failedReason', 'cost exceeds token bucket capacity',
+            'finishedOn', tostring(ts))
+          emitEvent(prefix .. 'events', 'failed', headJobId, {'failedReason', 'cost exceeds token bucket capacity'})
+          recordMetrics(metricsKey, ts, ts - processedOn)
+        elseif tbTokensCur < headCost then
+          -- Not enough tokens: register delay and skip promotion
+          local tbRateVal = tonumber(g.tbRefillRate) or 0
+          if tbRateVal <= 0 then break end
+          local tbDelayMs = math.ceil((headCost - tbTokensCur) * 1000 / tbRateVal)
+          local rateLimitedKey = prefix .. 'ratelimited'
+          redis.call('ZADD', rateLimitedKey, ts + tbDelayMs, gk)
+          return
+        else
+          tbOk = true
+          break
+        end
+      end
+    end
+    if not tbOk and tbCheckPasses >= 10 then return end
+  end
+  -- Calculate how many slots are available for promotion
+  local available = 1
+  if maxConc > 0 then
+    available = math.min(maxConc - newActive, 1000)
+  else
+    available = math.min(waitLen, 1000)
+  end
+  -- Cap by rate limit remaining if a window is active
+  if rateRemaining > 0 then
+    available = math.min(available, rateRemaining)
+  end
+  local streamKey = prefix .. 'stream'
+  local nextSeq = tonumber(g.nextSeq) or 0
+  if nextSeq > 0 then
+    local advanced = advancePastSkips(groupHashKey, nextSeq)
+    if advanced > nextSeq then
+      redis.call('HSET', groupHashKey, 'nextSeq', tostring(advanced))
+      nextSeq = advanced
+    end
+  end
+  local promoted = 0
+  local maxIter = available + 20
+  local iter = 0
+  while promoted < available and iter < maxIter do
+    iter = iter + 1
+    local zpResult = redis.call('ZPOPMIN', waitListKey, 1)
+    local nextJobId = zpResult[1]
+    if not nextJobId then break end
+    local nextJobKey = prefix .. 'job:' .. nextJobId
+    -- Skip stale entries (job no longer in group-waiting state)
+    local nextState = redis.call('HGET', nextJobKey, 'state')
+    if nextState ~= 'group-waiting' then
+      -- Stale: already processed via another path. Discard.
+    else
+      if nextSeq > 0 then
+        local jobSeq = tonumber(redis.call('HGET', nextJobKey, 'orderingSeq')) or 0
+        if jobSeq > nextSeq then
+          -- Future job: put back and stop promoting
+          redis.call('ZADD', waitListKey, jobSeq, nextJobId)
+          break
+        end
+      end
+      xaddJob(streamKey, nextJobId, redis.call('HGET', nextJobKey, 'name'))
+      redis.call('HSET', nextJobKey, 'state', 'waiting')
+      promoted = promoted + 1
+      -- Don't advance nextSeq here; moveToActive does it on actual activation
+      if nextSeq > 0 then nextSeq = nextSeq + 1 end -- local only, for stale skip in next iteration
+    end
+  end
+end
+
+local function expireJob(jobKey, jobId, prefix, now, curState, hintOrderingKey, hintOrderingSeq, hintGroupKey)
+  if curState == 'failed' then return true end
+  local wasActive = (curState == 'active')
+  local failedKey = prefix .. 'failed'
+  local eventsKey = prefix .. 'events'
+  local metricsKey = prefix .. 'metrics:failed'
+  local processedOn = tonumber(redis.call('HGET', jobKey, 'processedOn')) or now
+  redis.call('ZADD', failedKey, now, jobId)
+  redis.call('HSET', jobKey,
+    'state', 'failed',
+    'failedReason', 'expired',
+    'finishedOn', tostring(now))
+  markOrderingDone(jobKey, jobId, hintOrderingKey, hintOrderingSeq)
+  -- Only release group slot if the job was actually active (held a slot)
+  if wasActive then
+    releaseGroupSlotAndPromote(jobKey, jobId, now, hintGroupKey)
+  end
+  emitEvent(eventsKey, 'expired', jobId, nil)
+  recordMetrics(metricsKey, now, now - processedOn)
+  return true
+end
+
+local function checkExpired(jobKey, jobId, prefix, now)
+  local expireAt = tonumber(redis.call('HGET', jobKey, 'expireAt'))
+  if not expireAt or expireAt <= 0 then return false end
+  if now <= expireAt then return false end
+  -- Idempotency guard: if already expired, skip side effects
+  local curState = redis.call('HGET', jobKey, 'state')
+  return expireJob(jobKey, jobId, prefix, now, curState, nil, nil, nil)
+end
+
+local function advanceIdCounter(idKey, customId)
+  if not string.match(customId, '^%d+$') then return end
+  local numericId = tonumber(customId)
+  if numericId and numericId > 0 then
+    local cur = tonumber(redis.call('GET', idKey)) or 0
+    if numericId > cur then
+      redis.call('SET', idKey, customId)
+    end
+  end
+end
+
 local function extractOrderingKeyFromOpts(optsJson)
   if not optsJson or optsJson == '' then
     return ''
@@ -73,37 +354,142 @@ local function extractOrderingKeyFromOpts(optsJson)
   return tostring(key)
 end
 
--- glidemq_version()
--- Returns the library version string.
--- KEYS: (none)  ARGS: (none)
+local function extractLifoFromOpts(optsJson)
+  if not optsJson or optsJson == '' then return 0 end
+  local ok, decoded = pcall(cjson.decode, optsJson)
+  if not ok or type(decoded) ~= 'table' then return 0 end
+  return (decoded['lifo'] == true or decoded['lifo'] == 1) and 1 or 0
+end
+
+local function extractGroupConcurrencyFromOpts(optsJson)
+  if not optsJson or optsJson == '' then
+    return 0
+  end
+  local ok, decoded = pcall(cjson.decode, optsJson)
+  if not ok or type(decoded) ~= 'table' then
+    return 0
+  end
+  local ordering = decoded['ordering']
+  if type(ordering) ~= 'table' then
+    return 0
+  end
+  local conc = ordering['concurrency']
+  if conc == nil then
+    return 0
+  end
+  return tonumber(conc) or 0
+end
+
+local function extractGroupRateLimitFromOpts(optsJson)
+  if not optsJson or optsJson == '' then
+    return 0, 0
+  end
+  local ok, decoded = pcall(cjson.decode, optsJson)
+  if not ok or type(decoded) ~= 'table' then
+    return 0, 0
+  end
+  local ordering = decoded['ordering']
+  if type(ordering) ~= 'table' then
+    return 0, 0
+  end
+  local rl = ordering['rateLimit']
+  if type(rl) ~= 'table' then
+    return 0, 0
+  end
+  local max = tonumber(rl['max']) or 0
+  local duration = tonumber(rl['duration']) or 0
+  return max, duration
+end
+
+local function extractTokenBucketFromOpts(optsJson)
+  if not optsJson or optsJson == '' then return 0, 0 end
+  local ok, decoded = pcall(cjson.decode, optsJson)
+  if not ok or type(decoded) ~= 'table' then return 0, 0 end
+  local ordering = decoded['ordering']
+  if type(ordering) ~= 'table' then return 0, 0 end
+  local tb = ordering['tokenBucket']
+  if type(tb) ~= 'table' then return 0, 0 end
+  local capacity = tonumber(tb['capacity']) or 0
+  local refillRate = tonumber(tb['refillRate']) or 0
+  return math.floor(capacity * 1000), math.floor(refillRate * 1000)
+end
+
+local function extractCostFromOpts(optsJson)
+  if not optsJson or optsJson == '' then return 0 end
+  local ok, decoded = pcall(cjson.decode, optsJson)
+  if not ok or type(decoded) ~= 'table' then return 0 end
+  local cost = tonumber(decoded['cost']) or 0
+  return math.floor(cost * 1000)
+end
+
+local function extractTtlFromOpts(optsJson)
+  if not optsJson or optsJson == '' then return 0 end
+  local ok, decoded = pcall(cjson.decode, optsJson)
+  if not ok or type(decoded) ~= 'table' then return 0 end
+  return tonumber(decoded['ttl']) or 0
+end
+
+local function extractLockDurationFromOpts(optsJson)
+  if not optsJson or optsJson == '' then return 0 end
+  local ok, decoded = pcall(cjson.decode, optsJson)
+  if not ok or type(decoded) ~= 'table' then return 0 end
+  local lockDuration = tonumber(decoded['lockDuration']) or 0
+  -- Bounds must mirror MIN/MAX_JOB_LOCK_DURATION_MS in src/queue.ts and
+  -- MIN/MAX_JOB_OPTS_LOCK_DURATION_MS in src/proxy/routes.ts. Out-of-range
+  -- values are treated as unset so server-side reclaim falls back to the
+  -- worker lockDuration / minIdleMs.
+  if lockDuration < 1000 or lockDuration > 86400000 then return 0 end
+  return lockDuration
+end
+
+-- Apply stall logic to a job: increment stalledCount, fail if over max, else emit stalled event.
+-- Returns true if the job was moved to failed, false if only stalled.
+local function applyStalledLogic(jobKey, jobId, prefix, eventsKey, failedKey, maxStalledCount, timestamp)
+  local stalledCount = redis.call('HINCRBY', jobKey, 'stalledCount', 1)
+  if stalledCount > maxStalledCount then
+    local metricsKey = prefix .. 'metrics:failed'
+    local processedOn = tonumber(redis.call('HGET', jobKey, 'processedOn')) or timestamp
+    redis.call('ZADD', failedKey, timestamp, jobId)
+    redis.call('HSET', jobKey,
+      'state', 'failed',
+      'failedReason', 'job stalled more than maxStalledCount',
+      'finishedOn', tostring(timestamp)
+    )
+    markOrderingDone(jobKey, jobId)
+    releaseGroupSlotAndPromote(jobKey, jobId, timestamp)
+    emitEvent(eventsKey, 'failed', jobId, {
+      'failedReason', 'job stalled more than maxStalledCount'
+    })
+    recordMetrics(metricsKey, timestamp, timestamp - processedOn)
+    return true
+  else
+    emitEvent(eventsKey, 'stalled', jobId, nil)
+    return false
+  end
+end
+
+-- Remove excess jobs from a sorted set in capped, stack-safe batches.
+-- UNLINKs job hashes (they can be MB-sized with data + opts) and removes
+-- from the set in chunks of 1000.
+local function removeExcessJobs(setKey, prefix, ids)
+  for i = 1, #ids do
+    redis.call('UNLINK', prefix .. 'job:' .. ids[i])
+  end
+  for i = 1, #ids, 1000 do
+    redis.call('ZREM', setKey, unpack(ids, i, math.min(i + 999, #ids)))
+  end
+end
+
 redis.register_function('glidemq_version', function(keys, args)
-  return '1'
+  return '__LIBRARY_VERSION__'
 end)
 
--- glidemq_addJob(KEYS, ARGS)
--- KEYS[1] = id counter key
--- KEYS[2] = stream key
--- KEYS[3] = scheduled zset key
--- KEYS[4] = events stream key
--- (job hash key is derived from KEYS[1] prefix + ':job:' + id)
--- ARGS[1] = job name
--- ARGS[2] = job data (JSON string)
--- ARGS[3] = job opts (JSON string)
--- ARGS[4] = timestamp (ms)
--- ARGS[5] = delay (ms, 0 = no delay)
--- ARGS[6] = priority (0 = no priority)
--- ARGS[7] = parentId (empty string = no parent)
--- ARGS[8] = maxAttempts (0 = no retry)
--- ARGS[9] = orderingKey (empty string = no per-key ordering)
--- ARGS[10]-[16] = groupConcurrency, groupRateMax, groupRateDuration, tbCapacity, tbRefillRate, jobCost, ttl (see index.ts)
--- ARGS[17] = customJobId (empty string = auto-generate via INCR)
--- Returns: job ID (string), 'duplicate', or 'ERR:ID_EXHAUSTED'
 redis.register_function('glidemq_addJob', function(keys, args)
   local idKey = keys[1]
+  assert(string.sub(idKey, -3) == ':id', 'unexpected key format: ' .. idKey)
   local streamKey = keys[2]
   local scheduledKey = keys[3]
   local eventsKey = keys[4]
-
   local jobName = args[1]
   local jobData = args[2]
   local jobOpts = args[3]
@@ -113,42 +499,97 @@ redis.register_function('glidemq_addJob', function(keys, args)
   local parentId = args[7] or ''
   local maxAttempts = tonumber(args[8]) or 0
   local orderingKey = args[9] or ''
+  local groupConcurrency = tonumber(args[10]) or 0
+  local groupRateMax = tonumber(args[11]) or 0
+  local groupRateDuration = tonumber(args[12]) or 0
+  local tbCapacity = tonumber(args[13]) or 0
+  local tbRefillRate = tonumber(args[14]) or 0
+  local jobCost = tonumber(args[15]) or 0
+  local ttl = tonumber(args[16]) or 0
   local customJobId = args[17] or ''
-
-  -- Derive the job hash key from the id key prefix
-  -- idKey is like "glide:{q}:id", job key is "glide:{q}:job:{id}"
-  local prefix = string.sub(idKey, 1, #idKey - 2) -- strip ":id"
-
+  local lifo = tonumber(args[18]) or 0
+  local parentQueue = args[19] or ''
+  local schedulerName = args[20] or ''
+  local skipEvents = args[21] or '0'
+  local prefix = string.sub(idKey, 1, #idKey - 2)
   local jobIdStr
   local jobKey
   if customJobId ~= '' then
-    -- Custom job ID: check for duplicate
     jobKey = prefix .. 'job:' .. customJobId
     if redis.call('EXISTS', jobKey) == 1 then
       return 'duplicate'
     end
     jobIdStr = customJobId
+    advanceIdCounter(idKey, customJobId)
   else
-    -- Auto-generate via INCR, skip over IDs that collide with custom-ID jobs
     local jobId = redis.call('INCR', idKey)
     jobIdStr = tostring(jobId)
     jobKey = prefix .. 'job:' .. jobIdStr
-    local retries = 0
-    while redis.call('EXISTS', jobKey) == 1 do
-      retries = retries + 1
-      if retries >= 1000 then return 'ERR:ID_EXHAUSTED' end
-      jobId = redis.call('INCR', idKey)
-      jobIdStr = tostring(jobId)
-      jobKey = prefix .. 'job:' .. jobIdStr
-    end
   end
+  local useGroupConcurrency = (orderingKey ~= '')
   local orderingSeq = 0
   if orderingKey ~= '' then
     local orderingMetaKey = prefix .. 'ordering'
     orderingSeq = redis.call('HINCRBY', orderingMetaKey, orderingKey, 1)
   end
-
-  -- Store job hash
+  if useGroupConcurrency then
+    if groupConcurrency < 1 then groupConcurrency = 1 end
+    local groupHashKey = prefix .. 'group:' .. orderingKey
+    local curMax = tonumber(redis.call('HGET', groupHashKey, 'maxConcurrency')) or 0
+    if curMax ~= groupConcurrency then
+      redis.call('HSET', groupHashKey, 'maxConcurrency', tostring(groupConcurrency))
+    end
+    -- Initialize nextSeq for ordered promotion
+    if orderingSeq > 0 and redis.call('HEXISTS', groupHashKey, 'nextSeq') == 0 then
+      redis.call('HSET', groupHashKey, 'nextSeq', '1')
+    end
+    -- Upsert rate limit fields on group hash
+    if groupRateMax > 0 then
+      local curRateMax = tonumber(redis.call('HGET', groupHashKey, 'rateMax')) or 0
+      if curRateMax ~= groupRateMax then
+        redis.call('HSET', groupHashKey, 'rateMax', tostring(groupRateMax))
+      end
+      local curRateDuration = tonumber(redis.call('HGET', groupHashKey, 'rateDuration')) or 0
+      if curRateDuration ~= groupRateDuration then
+        redis.call('HSET', groupHashKey, 'rateDuration', tostring(groupRateDuration))
+      end
+    else
+      -- Clear stale rate limit fields if group was previously rate-limited
+      local oldRateMax = tonumber(redis.call('HGET', groupHashKey, 'rateMax')) or 0
+      if oldRateMax > 0 then
+        redis.call('HDEL', groupHashKey, 'rateMax', 'rateDuration', 'rateWindowStart', 'rateCount')
+      end
+    end
+    -- Upsert token bucket fields on group hash
+    if tbCapacity > 0 then
+      local curTbCap = tonumber(redis.call('HGET', groupHashKey, 'tbCapacity')) or 0
+      if curTbCap ~= tbCapacity then
+        redis.call('HSET', groupHashKey, 'tbCapacity', tostring(tbCapacity))
+      end
+      local curTbRate = tonumber(redis.call('HGET', groupHashKey, 'tbRefillRate')) or 0
+      if curTbRate ~= tbRefillRate then
+        redis.call('HSET', groupHashKey, 'tbRefillRate', tostring(tbRefillRate))
+      end
+      -- Initialize tokens on first setup
+      if curTbCap == 0 then
+        redis.call('HSET', groupHashKey,
+          'tbTokens', tostring(tbCapacity),
+          'tbLastRefill', tostring(timestamp),
+          'tbRefillRemainder', '0')
+      end
+      -- Validate cost (explicit or default 1000 millitokens) against capacity
+      local effectiveCost = (jobCost > 0) and jobCost or 1000
+      if effectiveCost > tbCapacity then
+        return 'ERR:COST_EXCEEDS_CAPACITY'
+      end
+    else
+      -- Clear stale tb fields
+      local oldTbCap = tonumber(redis.call('HGET', groupHashKey, 'tbCapacity')) or 0
+      if oldTbCap > 0 then
+        redis.call('HDEL', groupHashKey, 'tbCapacity', 'tbRefillRate', 'tbTokens', 'tbLastRefill', 'tbRefillRemainder')
+      end
+    end
+  end
   local hashFields = {
     'id', jobIdStr,
     'name', jobName,
@@ -160,19 +601,38 @@ redis.register_function('glidemq_addJob', function(keys, args)
     'priority', tostring(priority),
     'maxAttempts', tostring(maxAttempts)
   }
-
-  if orderingKey ~= '' then
-    hashFields[#hashFields + 1] = 'orderingKey'
+  if useGroupConcurrency then
+    hashFields[#hashFields + 1] = 'groupKey'
     hashFields[#hashFields + 1] = orderingKey
-    hashFields[#hashFields + 1] = 'orderingSeq'
-    hashFields[#hashFields + 1] = tostring(orderingSeq)
+    if orderingSeq > 0 then
+      hashFields[#hashFields + 1] = 'orderingSeq'
+      hashFields[#hashFields + 1] = tostring(orderingSeq)
+    end
   end
-
+  if jobCost > 0 then
+    hashFields[#hashFields + 1] = 'cost'
+    hashFields[#hashFields + 1] = tostring(jobCost)
+  end
+  if ttl > 0 then
+    hashFields[#hashFields + 1] = 'expireAt'
+    hashFields[#hashFields + 1] = tostring(timestamp + ttl)
+  end
   if parentId ~= '' then
     hashFields[#hashFields + 1] = 'parentId'
     hashFields[#hashFields + 1] = parentId
+    if parentQueue ~= '' then
+      hashFields[#hashFields + 1] = 'parentQueue'
+      hashFields[#hashFields + 1] = parentQueue
+    end
   end
-
+  if schedulerName ~= '' then
+    hashFields[#hashFields + 1] = 'schedulerName'
+    hashFields[#hashFields + 1] = schedulerName
+  end
+  if lifo > 0 then
+    hashFields[#hashFields + 1] = 'lifo'
+    hashFields[#hashFields + 1] = '1'
+  end
   if delay > 0 or priority > 0 then
     hashFields[#hashFields + 1] = 'state'
     hashFields[#hashFields + 1] = delay > 0 and 'delayed' or 'prioritized'
@@ -180,48 +640,37 @@ redis.register_function('glidemq_addJob', function(keys, args)
     hashFields[#hashFields + 1] = 'state'
     hashFields[#hashFields + 1] = 'waiting'
   end
-
   redis.call('HSET', jobKey, unpack(hashFields))
-
-  -- Route job: delayed/prioritized go to scheduled ZSet, otherwise to stream
+  -- Register child in parent's deps set when parentDepsKey is provided (keys[5])
+  if parentId ~= '' and parentQueue ~= '' and #keys >= 5 then
+    local parentDepsKey = keys[5]
+    -- prefix includes trailing colon (glide:{Q}:), so strip it for depsMember
+    local queuePrefix = string.sub(prefix, 1, #prefix - 1)
+    local depsMember = queuePrefix .. ':' .. jobIdStr
+    redis.call('SADD', parentDepsKey, depsMember)
+  end
   if delay > 0 then
-    -- Delayed: score = (priority * 2^42) + (timestamp + delay)
     local score = priority * PRIORITY_SHIFT + (timestamp + delay)
     redis.call('ZADD', scheduledKey, score, jobIdStr)
   elseif priority > 0 then
-    -- Prioritized but not delayed: score = (priority * 2^42) + 0
-    -- timestamp=0 so it promotes immediately on next promote cycle
     local score = priority * PRIORITY_SHIFT
     redis.call('ZADD', scheduledKey, score, jobIdStr)
+  elseif lifo > 0 then
+    local lifoKey = prefix .. 'lifo'
+    redis.call('RPUSH', lifoKey, jobIdStr)
   else
-    -- Immediate: add to stream
-    redis.call('XADD', streamKey, '*', 'jobId', jobIdStr)
+    xaddJob(streamKey, jobIdStr, jobName)
   end
-
-  -- Emit 'added' event
-  emitEvent(eventsKey, 'added', jobIdStr, {'name', jobName})
-
+  if skipEvents ~= '1' then emitEvent(eventsKey, 'added', jobIdStr, {'name', jobName}) end
   return jobIdStr
 end)
 
--- glidemq_promote(KEYS, ARGS)
--- Promotes jobs from the scheduled ZSet to the stream when their time has come.
--- KEYS[1] = scheduled zset key
--- KEYS[2] = stream key
--- KEYS[3] = events stream key
--- ARGS[1] = current timestamp (ms)
--- Returns: number of promoted jobs
 redis.register_function('glidemq_promote', function(keys, args)
   local scheduledKey = keys[1]
   local streamKey = keys[2]
   local eventsKey = keys[3]
-
   local now = tonumber(args[1])
   local MAX_PROMOTIONS = 1000
-
-  -- Scores are encoded as: (priority * PRIORITY_SHIFT) + scheduledAt.
-  -- Iterate only populated priority bands to avoid full scans and sparse huge loops.
-
   local count = 0
   local cursorMin = 0
   while count < MAX_PROMOTIONS do
@@ -234,57 +683,126 @@ redis.register_function('glidemq_promote', function(keys, args)
     local minScore = priority * PRIORITY_SHIFT
     local maxDueScore = minScore + now
     local remaining = MAX_PROMOTIONS - count
-    local members = redis.call('ZRANGEBYSCORE', scheduledKey, string.format('%.0f', minScore), string.format('%.0f', maxDueScore), 'LIMIT', 0, remaining)
+    local members = redis.call(
+      'ZRANGEBYSCORE',
+      scheduledKey,
+      string.format('%.0f', minScore),
+      string.format('%.0f', maxDueScore),
+      'LIMIT',
+      0,
+      remaining
+    )
     for i = 1, #members do
-      local jobId = members[i]
-      redis.call('XADD', streamKey, '*', 'jobId', jobId)
-      redis.call('ZREM', scheduledKey, jobId)
-
-      -- Update job state
-      -- Derive the prefix from scheduledKey: "glide:{q}:scheduled" -> "glide:{q}:"
-      local prefix = string.sub(scheduledKey, 1, #scheduledKey - 9) -- strip "scheduled"
+      local member = members[i]
+      local sepStart, sepEnd = string.find(member, '||', 1, true)
+      local jobId = member
+      local retryGroup = nil
+      if sepStart then
+        jobId = string.sub(member, 1, sepStart - 1)
+        retryGroup = string.sub(member, sepEnd + 1)
+        if retryGroup == '' then retryGroup = nil end
+      end
+      local prefix = string.sub(scheduledKey, 1, #scheduledKey - 9)
       local jobKey = prefix .. 'job:' .. jobId
-      redis.call('HSET', jobKey, 'state', 'waiting')
-
-      emitEvent(eventsKey, 'promoted', jobId, nil)
+      redis.call('ZREM', scheduledKey, member)
       count = count + 1
+      if not checkExpired(jobKey, jobId, prefix, now) then
+        local jobLifo = redis.call('HGET', jobKey, 'lifo')
+        if jobLifo == '1' then
+          local lifoKey = prefix .. 'lifo'
+          redis.call('RPUSH', lifoKey, jobId)
+        elseif priority > 0 then
+          -- Priority jobs go to dedicated priority list, checked before LIFO and stream.
+          -- LPUSH so RPOP retrieves lowest-score (highest-priority) job first.
+          local priorityKey = prefix .. 'priority'
+          redis.call('LPUSH', priorityKey, jobId)
+        else
+          local jobName = redis.call('HGET', jobKey, 'name')
+          if retryGroup then
+            redis.call('XADD', streamKey, '*', 'jobId', jobId, 'name', jobName or '', 'retryGroup', retryGroup)
+          else
+            xaddJob(streamKey, jobId, jobName)
+          end
+        end
+        redis.call('HSET', jobKey, 'state', 'waiting')
+        emitEvent(eventsKey, 'promoted', jobId, nil)
+      end
+      if count >= MAX_PROMOTIONS then break end
     end
     cursorMin = (priority + 1) * PRIORITY_SHIFT
   end
-
   return count
 end)
 
--- glidemq_complete(KEYS, ARGS)
--- Moves a job from active (PEL) to completed.
--- After completion, if the job has a parent, triggers completeChild logic.
--- KEYS[1] = stream key
--- KEYS[2] = completed zset key
--- KEYS[3] = events stream key
--- KEYS[4] = job hash key (job:{id})
--- Optional parent keys (when job has parentId):
--- KEYS[5] = parent deps key (glide:{parentQueue}:deps:{parentId})
--- KEYS[6] = parent job hash key (glide:{parentQueue}:job:{parentId})
--- KEYS[7] = parent stream key (glide:{parentQueue}:stream)
--- KEYS[8] = parent events key (glide:{parentQueue}:events)
--- ARGS[1] = job ID
--- ARGS[2] = stream entry ID (for XACK)
--- ARGS[3] = return value (JSON string)
--- ARGS[4] = timestamp (ms)
--- ARGS[5] = consumer group name
--- ARGS[6] = removeMode ('0'|'true'|'count'|'age_count')
--- ARGS[7] = removeCount (number)
--- ARGS[8] = removeAge (seconds)
--- ARGS[9] = depsMember (childQueuePrefix:childId, empty string = no parent)
--- ARGS[10] = parentId (empty string = no parent)
--- Returns: 1 on success
+redis.register_function('glidemq_nextDue', function(keys, args)
+  local scheduledKey = keys[1]
+  local rateLimitedKey = keys[2]
+  local nextDue = nil
+
+  local scheduled = redis.call('ZRANGE', scheduledKey, 0, 0, 'WITHSCORES')
+  if scheduled and #scheduled >= 2 then
+    local score = tonumber(scheduled[2]) or 0
+    local due = score % PRIORITY_SHIFT
+    nextDue = due
+  end
+
+  local limited = redis.call('ZRANGE', rateLimitedKey, 0, 0, 'WITHSCORES')
+  if limited and #limited >= 2 then
+    local limitedDue = tonumber(limited[2]) or 0
+    if (not nextDue) or limitedDue < nextDue then
+      nextDue = limitedDue
+    end
+  end
+
+  if not nextDue then
+    return -1
+  end
+
+  return math.floor(nextDue)
+end)
+
+redis.register_function('glidemq_tryLock', function(keys, args)
+  local lockKey = keys[1]
+  local token = args[1]
+  local ttl = tonumber(args[2]) or 1000
+  local result = redis.call('SET', lockKey, token, 'PX', tostring(ttl), 'NX')
+  if result then
+    return 1
+  end
+  return 0
+end)
+
+redis.register_function('glidemq_unlock', function(keys, args)
+  local lockKey = keys[1]
+  local token = args[1]
+  local current = redis.call('GET', lockKey)
+  if current == token then
+    redis.call('DEL', lockKey)
+    return 1
+  end
+  return 0
+end)
+
+redis.register_function('glidemq_renewLock', function(keys, args)
+  local lockKey = keys[1]
+  local token = args[1]
+  local ttl = tonumber(args[2]) or 1000
+  local current = redis.call('GET', lockKey)
+  if current == token then
+    redis.call('PEXPIRE', lockKey, ttl)
+    return 1
+  end
+  return 0
+end)
+
 redis.register_function('glidemq_complete', function(keys, args)
   local streamKey = keys[1]
   local completedKey = keys[2]
   local eventsKey = keys[3]
   local jobKey = keys[4]
-
+  local metricsKey = keys[5]
   local jobId = args[1]
+  assert(string.sub(jobKey, -(4 + #jobId)) == 'job:' .. jobId, 'unexpected key format: ' .. jobKey)
   local entryId = args[2]
   local returnvalue = args[3]
   local timestamp = tonumber(args[4])
@@ -294,114 +812,547 @@ redis.register_function('glidemq_complete', function(keys, args)
   local removeAge = tonumber(args[8]) or 0
   local depsMember = args[9] or ''
   local parentId = args[10] or ''
-
-  -- Acknowledge the stream entry
-  redis.call('XACK', streamKey, group, entryId)
-
-  -- Remove the message from the stream (we no longer need it)
-  redis.call('XDEL', streamKey, entryId)
-
-  -- Add to completed ZSet (score = timestamp for ordering)
+  local broadcastMode = args[11] or '0'
+  local processedOn = tonumber(redis.call('HGET', jobKey, 'processedOn')) or timestamp
+  if entryId ~= '' then redis.call('XACK', streamKey, group, entryId) end
+  if entryId ~= '' and broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
   redis.call('ZADD', completedKey, timestamp, jobId)
-
-  -- Update job hash
   redis.call('HSET', jobKey,
     'state', 'completed',
     'returnvalue', returnvalue,
     'finishedOn', tostring(timestamp)
   )
   markOrderingDone(jobKey, jobId)
-
-  -- Emit event
+  releaseGroupSlotAndPromote(jobKey, jobId, timestamp)
   emitEvent(eventsKey, 'completed', jobId, {'returnvalue', returnvalue})
-
-  -- Retention cleanup
+  recordMetrics(metricsKey, timestamp, timestamp - processedOn)
   local prefix = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
-  if removeMode == 'true' then
-    redis.call('ZREM', completedKey, jobId)
-    redis.call('UNLINK', jobKey)
-  elseif removeMode == 'count' and removeCount > 0 then
-    local total = redis.call('ZCARD', completedKey)
-    if total > removeCount then
-      local excess = redis.call('ZRANGE', completedKey, 0, total - removeCount - 1)
-      for i = 1, #excess do
-        local oldId = excess[i]
-        redis.call('UNLINK', prefix .. 'job:' .. oldId)
-        redis.call('ZREM', completedKey, oldId)
-      end
-    end
-  elseif removeMode == 'age_count' then
-    if removeAge > 0 then
-      local cutoff = timestamp - (removeAge * 1000)
-      local old = redis.call('ZRANGEBYSCORE', completedKey, '0', tostring(cutoff))
-      for i = 1, #old do
-        local oldId = old[i]
-        redis.call('UNLINK', prefix .. 'job:' .. oldId)
-        redis.call('ZREM', completedKey, oldId)
-      end
-    end
-    if removeCount > 0 then
+  if broadcastMode ~= '1' then
+    if removeMode == 'true' then
+      redis.call('ZREM', completedKey, jobId)
+      redis.call('UNLINK', jobKey)
+    elseif removeMode == 'count' and removeCount > 0 then
       local total = redis.call('ZCARD', completedKey)
       if total > removeCount then
-        local excess = redis.call('ZRANGE', completedKey, 0, total - removeCount - 1)
-        for i = 1, #excess do
-          local oldId = excess[i]
-          redis.call('UNLINK', prefix .. 'job:' .. oldId)
-          redis.call('ZREM', completedKey, oldId)
+        local excess = redis.call('ZRANGE', completedKey, 0, math.min(total - removeCount, 1000) - 1)
+        if #excess > 0 then removeExcessJobs(completedKey, prefix, excess) end
+      end
+    elseif removeMode == 'age_count' then
+      if removeAge > 0 then
+        local cutoff = timestamp - (removeAge * 1000)
+        local old = redis.call('ZRANGEBYSCORE', completedKey, '0', string.format('%.0f', cutoff), 'LIMIT', 0, 1000)
+        if #old > 0 then removeExcessJobs(completedKey, prefix, old) end
+      end
+      if removeCount > 0 then
+        local total = redis.call('ZCARD', completedKey)
+        if total > removeCount then
+          local excess = redis.call('ZRANGE', completedKey, 0, math.min(total - removeCount, 1000) - 1)
+          if #excess > 0 then removeExcessJobs(completedKey, prefix, excess) end
+        end
+      end
+    end
+  end
+  if depsMember ~= '' and parentId ~= '' and #keys >= 9 then
+    local parentDepsKey = keys[6]
+    local parentJobKey = keys[7]
+    local parentStreamKey = keys[8]
+    local parentEventsKey = keys[9]
+    local depMarker = 'depdone:' .. depsMember
+    if redis.call('HSETNX', parentJobKey, depMarker, '1') == 1 then
+      local doneCount = redis.call('HINCRBY', parentJobKey, 'depsCompleted', 1)
+      local totalDeps = redis.call('SCARD', parentDepsKey)
+      local remaining = totalDeps - doneCount
+      if remaining <= 0 then
+        local parentState = redis.call('HGET', parentJobKey, 'state')
+        if parentState == 'waiting-children' then
+          redis.call('HSET', parentJobKey, 'state', 'waiting')
+          xaddJob(parentStreamKey, parentId, redis.call('HGET', parentJobKey, 'name'))
+          emitEvent(parentEventsKey, 'active', parentId, nil)
+        end
+      end
+    end
+  end
+  -- DAG multi-parent: notify additional same-queue parents via parents SET
+  local parentsKey = prefix .. 'parents:' .. jobId
+  local dagParents = redis.call('SMEMBERS', parentsKey)
+  if dagParents and #dagParents > 0 then
+    local childQueuePrefix = string.sub(prefix, 1, #prefix - 1)
+    local dagDepsMember = childQueuePrefix .. ':' .. jobId
+    for pi = 1, #dagParents do
+      local pEntry = dagParents[pi]
+      -- Format: "parentQueuePrefix:parentId" where prefix is glide:{qname}
+      -- Must find LAST colon (not first, since prefix contains colons in {})
+      local pSep = pEntry:find(':([^:]+)$')
+      if pSep then
+        local pQueue = string.sub(pEntry, 1, pSep - 1)
+        local pId = string.sub(pEntry, pSep + 1)
+        -- Only handle same-queue parents atomically in Lua.
+        -- Cross-queue parents are skipped here because their keys use a different
+        -- hash tag (different prefix), which may route to a different cluster slot.
+        -- The TypeScript layer handles cross-queue parent notification separately.
+        if pQueue == childQueuePrefix then
+          local pPrefix = prefix
+          local pJobKey = pPrefix .. 'job:' .. pId
+          local pDepsKey = pPrefix .. 'deps:' .. pId
+          local pStreamKey = pPrefix .. 'stream'
+          local pEventsKey = pPrefix .. 'events'
+          local pDepMarker = 'depdone:' .. dagDepsMember
+          if redis.call('HSETNX', pJobKey, pDepMarker, '1') == 1 then
+            local pDoneCount = redis.call('HINCRBY', pJobKey, 'depsCompleted', 1)
+            local pTotalDeps = redis.call('SCARD', pDepsKey)
+            if pTotalDeps - pDoneCount <= 0 then
+              local pState = redis.call('HGET', pJobKey, 'state')
+              if pState == 'waiting-children' then
+                redis.call('HSET', pJobKey, 'state', 'waiting')
+                xaddJob(pStreamKey, pId, redis.call('HGET', pJobKey, 'name'))
+                emitEvent(pEventsKey, 'active', pId, nil)
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+  if entryId == '' then decrListActive(prefix .. 'list-active') end
+  return 1
+end)
+
+redis.register_function('glidemq_completeAndFetchNext', function(keys, args)
+  local streamKey = keys[1]
+  local completedKey = keys[2]
+  local eventsKey = keys[3]
+  local jobKey = keys[4]
+  local metricsKey = keys[5]
+  local jobId = args[1]
+  local entryId = args[2]
+  local returnvalue = args[3]
+  local timestamp = tonumber(args[4])
+  local group = args[5]
+  local consumer = args[6]
+  local removeMode = args[7] or '0'
+  local removeCount = tonumber(args[8]) or 0
+  local removeAge = tonumber(args[9]) or 0
+  local depsMember = args[10] or ''
+  local parentId = args[11] or ''
+  local currentOrderingKey = args[12] or ''
+  local currentOrderingSeq = args[13] or ''
+  local currentGroupKey = args[14] or ''
+  local broadcastMode = args[15] or '0'
+  local hintProcessedOn = args[16] or ''
+  local hasParents = args[17] or '0'
+  local skipEvents = args[18] or '0'
+  local skipMetrics = args[19] or '0'
+
+  -- Phase 1: Complete current job (same as glidemq_complete)
+  local processedOn
+  if hintProcessedOn ~= '' then
+    processedOn = tonumber(hintProcessedOn) or timestamp
+  else
+    processedOn = tonumber(redis.call('HGET', jobKey, 'processedOn')) or timestamp
+  end
+  if entryId ~= '' then redis.call('XACK', streamKey, group, entryId) end
+  if entryId ~= '' and broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
+  redis.call('ZADD', completedKey, timestamp, jobId)
+  redis.call('HSET', jobKey,
+    'state', 'completed',
+    'returnvalue', returnvalue,
+    'finishedOn', tostring(timestamp)
+  )
+  if currentOrderingKey ~= '__' then
+    markOrderingDone(jobKey, jobId, currentOrderingKey, currentOrderingSeq)
+  end
+  if currentGroupKey ~= '__' then
+    releaseGroupSlotAndPromote(jobKey, jobId, timestamp, currentGroupKey)
+  end
+  if skipEvents ~= '1' then emitEvent(eventsKey, 'completed', jobId, {'returnvalue', returnvalue}) end
+  if skipMetrics ~= '1' then recordMetrics(metricsKey, timestamp, timestamp - processedOn) end
+  local prefix = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
+  if entryId == '' then decrListActive(prefix .. 'list-active') end
+
+  -- Retention cleanup (skip in broadcast mode - job hash must persist for all subscriptions)
+  if broadcastMode ~= '1' then
+    if removeMode == 'true' then
+      redis.call('ZREM', completedKey, jobId)
+      redis.call('UNLINK', jobKey)
+    elseif removeMode == 'count' and removeCount > 0 then
+      local total = redis.call('ZCARD', completedKey)
+      if total > removeCount then
+        local excess = redis.call('ZRANGE', completedKey, 0, math.min(total - removeCount, 1000) - 1)
+        if #excess > 0 then removeExcessJobs(completedKey, prefix, excess) end
+      end
+    elseif removeMode == 'age_count' then
+      if removeAge > 0 then
+        local cutoff = timestamp - (removeAge * 1000)
+        local old = redis.call('ZRANGEBYSCORE', completedKey, '0', string.format('%.0f', cutoff), 'LIMIT', 0, 1000)
+        if #old > 0 then removeExcessJobs(completedKey, prefix, old) end
+      end
+      if removeCount > 0 then
+        local total = redis.call('ZCARD', completedKey)
+        if total > removeCount then
+          local excess = redis.call('ZRANGE', completedKey, 0, math.min(total - removeCount, 1000) - 1)
+          if #excess > 0 then removeExcessJobs(completedKey, prefix, excess) end
         end
       end
     end
   end
 
-  -- If this job has a parent, handle completeChild inline
-  if depsMember ~= '' and parentId ~= '' and #keys >= 8 then
-    local parentDepsKey = keys[5]
-    local parentJobKey = keys[6]
-    local parentStreamKey = keys[7]
-    local parentEventsKey = keys[8]
-
-    -- Increment completed children counter on parent hash
-    local doneCount = redis.call('HINCRBY', parentJobKey, 'depsCompleted', 1)
-    local totalDeps = redis.call('SCARD', parentDepsKey)
-    local remaining = totalDeps - doneCount
-
-    if remaining <= 0 then
-      -- All children done: move parent to stream
-      redis.call('HSET', parentJobKey, 'state', 'waiting')
-      redis.call('XADD', parentStreamKey, '*', 'jobId', parentId)
-      emitEvent(parentEventsKey, 'active', parentId, nil)
+  -- Parent deps
+  if depsMember ~= '' and parentId ~= '' and #keys >= 9 then
+    local parentDepsKey = keys[6]
+    local parentJobKey = keys[7]
+    local parentStreamKey = keys[8]
+    local parentEventsKey = keys[9]
+    local depMarker = 'depdone:' .. depsMember
+    if redis.call('HSETNX', parentJobKey, depMarker, '1') == 1 then
+      local doneCount = redis.call('HINCRBY', parentJobKey, 'depsCompleted', 1)
+      local totalDeps = redis.call('SCARD', parentDepsKey)
+      if totalDeps - doneCount <= 0 then
+        local parentState = redis.call('HGET', parentJobKey, 'state')
+        if parentState == 'waiting-children' then
+          redis.call('HSET', parentJobKey, 'state', 'waiting')
+          xaddJob(parentStreamKey, parentId, redis.call('HGET', parentJobKey, 'name'))
+          emitEvent(parentEventsKey, 'active', parentId, nil)
+        end
+      end
+    end
+  end
+  -- DAG multi-parent: always check parents SET. The previous hasParents
+  -- arg was sourced from the worker's snapshot of the job hash, which is
+  -- stale if registerParent populates the SET after the worker fetched the
+  -- job but before completion (race seen on the addDAG path). SMEMBERS on
+  -- a non-existent / empty SET costs ~nothing on the server, so always run.
+  do
+    local parentsKey = prefix .. 'parents:' .. jobId
+    local dagParents = redis.call('SMEMBERS', parentsKey)
+    if dagParents and #dagParents > 0 then
+      local childQueuePrefix = string.sub(prefix, 1, #prefix - 1)
+      local dagDepsMember = childQueuePrefix .. ':' .. jobId
+      for pi = 1, #dagParents do
+        local pEntry = dagParents[pi]
+        local pSep = pEntry:find(':([^:]+)$')
+        if pSep then
+          local pQueue = string.sub(pEntry, 1, pSep - 1)
+          local pId = string.sub(pEntry, pSep + 1)
+          if pQueue == childQueuePrefix then
+            local pPrefix = prefix
+            local pJobKey = pPrefix .. 'job:' .. pId
+            local pDepsKey = pPrefix .. 'deps:' .. pId
+            local pStreamKey = pPrefix .. 'stream'
+            local pEventsKey = pPrefix .. 'events'
+            local pDepMarker = 'depdone:' .. dagDepsMember
+            if redis.call('HSETNX', pJobKey, pDepMarker, '1') == 1 then
+              local pDoneCount = redis.call('HINCRBY', pJobKey, 'depsCompleted', 1)
+              local pTotalDeps = redis.call('SCARD', pDepsKey)
+              if pTotalDeps - pDoneCount <= 0 then
+                local pState = redis.call('HGET', pJobKey, 'state')
+                if pState == 'waiting-children' then
+                  redis.call('HSET', pJobKey, 'state', 'waiting')
+                  xaddJob(pStreamKey, pId, redis.call('HGET', pJobKey, 'name'))
+                  emitEvent(pEventsKey, 'active', pId, nil)
+                end
+              end
+            end
+          end
+        end
+      end
     end
   end
 
-  return 1
+  -- In broadcast mode: do not fetch next (avoids XDEL of next entry which would break other consumer groups)
+  if broadcastMode == '1' then
+    return {'NEXT_NONE', jobId}
+  end
+
+  -- Return protocol (array-based to avoid cjson encode/decode per job):
+  -- {'NEXT_NONE', completedJobId}
+  -- {'NEXT_REVOKED', completedJobId, nextJobId, nextEntryId}
+  -- {'NEXT_HASH', completedJobId, nextJobId, nextEntryId, field1, value1, field2, value2, ...}
+
+  -- Phase 1.0: Try priority list first (highest priority: priority > LIFO > FIFO)
+  local priorityKey = prefix .. 'priority'
+  for _priAttempt = 1, 3 do
+    local priJobId = redis.call('RPOP', priorityKey)
+    if not priJobId then
+      break  -- priority list is empty
+    end
+
+    local priJobKey = prefix .. 'job:' .. priJobId
+    -- Single HMGET replaces EXISTS + HGET 'revoked' + checkExpired HGET 'expireAt' (3 → 1)
+    local priMeta = redis.call('HMGET', priJobKey, 'state', 'revoked', 'expireAt')
+    if priMeta[1] then
+      if priMeta[2] ~= '1' then
+        local priExpireAt = tonumber(priMeta[3])
+        if not priExpireAt or priExpireAt <= 0 or timestamp <= priExpireAt then
+          -- Check ordering/group gates for priority-list jobs
+          local priGroupKey = redis.call('HGET', priJobKey, 'groupKey')
+          if priGroupKey and priGroupKey ~= '' then
+            local priGroupHashKey = prefix .. 'group:' .. priGroupKey
+            local priGrpFields = redis.call('HGETALL', priGroupHashKey)
+            local priGrp = {}
+            for pf = 1, #priGrpFields, 2 do priGrp[priGrpFields[pf]] = priGrpFields[pf + 1] end
+            local priOrdSeq = tonumber(redis.call('HGET', priJobKey, 'orderingSeq')) or 0
+            local priNextSeq = tonumber(priGrp.nextSeq) or 0
+            if priNextSeq > 0 then
+              local priAdv = advancePastSkips(priGroupHashKey, priNextSeq)
+              if priAdv > priNextSeq then
+                redis.call('HSET', priGroupHashKey, 'nextSeq', tostring(priAdv))
+                priNextSeq = priAdv
+              end
+            end
+            local priMaxConc = tonumber(priGrp.maxConcurrency) or 0
+            local priActive = tonumber(priGrp.active) or 0
+            local priWaitListKey = prefix .. 'groupq:' .. priGroupKey
+            local priReturning = (priOrdSeq > 0 and priNextSeq > 0 and priOrdSeq < priNextSeq)
+            -- Ordering gate or concurrency gate (skip for returning step-jobs)
+            if (priOrdSeq > 0 and priNextSeq > 0 and priOrdSeq > priNextSeq) or
+               (priMaxConc > 0 and priActive >= priMaxConc and not priReturning) then
+              local priScore = priOrdSeq > 0 and priOrdSeq or tonumber(priJobId) or 0
+              redis.call('ZADD', priWaitListKey, priScore, priJobId)
+              redis.call('HSET', priJobKey, 'state', 'group-waiting')
+            else
+              -- All gates pass: activate with group bookkeeping
+              redis.call('HSET', priJobKey, 'state', 'active', 'processedOn', tostring(timestamp), 'lastActive', tostring(timestamp))
+              if not priReturning then
+                redis.call('HINCRBY', priGroupHashKey, 'active', 1)
+              end
+              if priOrdSeq > 0 and not priReturning then
+                redis.call('HSET', priGroupHashKey, 'nextSeq', tostring(priOrdSeq + 1))
+                redis.call('ZREM', priWaitListKey, priJobId)
+              end
+              if skipEvents ~= '1' then emitEvent(eventsKey, 'active', priJobId, nil) end
+              local priJobFields = redis.call('HGETALL', priJobKey)
+              redis.call('INCR', prefix .. 'list-active')
+              return {'NEXT_HASH', jobId, priJobId, '', unpack(priJobFields)}
+            end
+          else
+            -- Non-group job: activate directly
+            redis.call('HSET', priJobKey, 'state', 'active', 'processedOn', tostring(timestamp), 'lastActive', tostring(timestamp))
+            if skipEvents ~= '1' then emitEvent(eventsKey, 'active', priJobId, nil) end
+            local priJobFields = redis.call('HGETALL', priJobKey)
+            redis.call('INCR', prefix .. 'list-active')
+            return {'NEXT_HASH', jobId, priJobId, '', unpack(priJobFields)}
+          end
+        else
+          expireJob(priJobKey, priJobId, prefix, timestamp, priMeta[1], nil, nil, nil)
+        end
+      end
+    end
+  end
+
+  -- Phase 1.5: Try LIFO list (before stream), retry up to 3 times
+  local lifoKey = prefix .. 'lifo'
+  for _lifoAttempt = 1, 3 do
+    local lifoJobId = redis.call('RPOP', lifoKey)
+    if not lifoJobId then
+      break  -- LIFO list is empty
+    end
+
+    local lifoJobKey = prefix .. 'job:' .. lifoJobId
+    -- Single HMGET replaces EXISTS + HGET 'revoked' + checkExpired HGET 'expireAt' (3 → 1)
+    local lifoMeta = redis.call('HMGET', lifoJobKey, 'state', 'revoked', 'expireAt')
+    if lifoMeta[1] then
+      if lifoMeta[2] ~= '1' then
+        local lifoExpireAt = tonumber(lifoMeta[3])
+        if not lifoExpireAt or lifoExpireAt <= 0 or timestamp <= lifoExpireAt then
+          redis.call('HSET', lifoJobKey, 'state', 'active', 'processedOn', tostring(timestamp), 'lastActive', tostring(timestamp))
+          if skipEvents ~= '1' then emitEvent(eventsKey, 'active', lifoJobId, nil) end
+          local lifoJobFields = redis.call('HGETALL', lifoJobKey)
+          redis.call('INCR', prefix .. 'list-active')
+          return {'NEXT_HASH', jobId, lifoJobId, '', unpack(lifoJobFields)}
+        else
+          expireJob(lifoJobKey, lifoJobId, prefix, timestamp, lifoMeta[1], nil, nil, nil)
+        end
+      end
+    end
+  end
+
+  -- Phase 2: Fetch next job (non-blocking XREADGROUP), skip expired (up to 3 attempts)
+  local nextJobId, nextEntryId, nextJobKey
+  local nextGroupKey = nil
+  for _fetchAttempt = 1, 3 do
+    local nextEntries = redis.call('XREADGROUP', 'GROUP', group, consumer, 'COUNT', 1, 'STREAMS', streamKey, '>')
+    if not nextEntries or #nextEntries == 0 then
+      return {'NEXT_NONE', jobId}
+    end
+    local streamData = nextEntries[1]
+    local entries = streamData[2]
+    if not entries or #entries == 0 then
+      return {'NEXT_NONE', jobId}
+    end
+    local nextEntry = entries[1]
+    nextEntryId = nextEntry[1]
+    local nextFields = nextEntry[2]
+    nextJobId = nil
+    for i = 1, #nextFields, 2 do
+      if nextFields[i] == 'jobId' then
+        nextJobId = nextFields[i + 1]
+        break
+      end
+    end
+    if not nextJobId then
+      return {'NEXT_NONE', jobId}
+    end
+    nextJobKey = prefix .. 'job:' .. nextJobId
+    -- Single HMGET replaces EXISTS + HGET 'revoked' + checkExpired HGET 'expireAt' + HGET 'groupKey' (4 → 1)
+    local nextMeta = redis.call('HMGET', nextJobKey, 'state', 'revoked', 'expireAt', 'groupKey')
+    if not nextMeta[1] then
+      -- state is nil: job hash does not exist
+      return {'NEXT_NONE', jobId}
+    end
+    if nextMeta[2] == '1' then
+      return {'NEXT_REVOKED', jobId, nextJobId, nextEntryId}
+    end
+    -- Inline expiry check (avoids checkExpired's redundant HGET)
+    local nextExpireAt = tonumber(nextMeta[3])
+    if nextExpireAt and nextExpireAt > 0 and timestamp > nextExpireAt then
+      local curState = nextMeta[1]
+      expireJob(nextJobKey, nextJobId, prefix, timestamp, curState, nil, nil, nil)
+      redis.call('XACK', streamKey, group, nextEntryId)
+      redis.call('XDEL', streamKey, nextEntryId)
+      nextJobId = nil
+    else
+      nextGroupKey = nextMeta[4]
+      break
+    end
+  end
+  if not nextJobId then
+    return {'NEXT_NONE', jobId}
+  end
+
+  -- Phase 3: Activate next job (same as moveToActive)
+  if nextGroupKey and nextGroupKey ~= '' then
+    local nextGroupHashKey = prefix .. 'group:' .. nextGroupKey
+    -- Load all group fields in one call
+    local nGrpFields = redis.call('HGETALL', nextGroupHashKey)
+    local nGrp = {}
+    for nf = 1, #nGrpFields, 2 do nGrp[nGrpFields[nf]] = nGrpFields[nf + 1] end
+    local nextMaxConc = tonumber(nGrp.maxConcurrency) or 0
+    local nextActive = tonumber(nGrp.active) or 0
+    local nextWaitListKey = prefix .. 'groupq:' .. nextGroupKey
+    -- Ordering gate
+    local nextJobOrderingSeq = tonumber(redis.call('HGET', nextJobKey, 'orderingSeq')) or 0
+    local nextReturning = false
+    if nextJobOrderingSeq > 0 then
+      local nextExpectedSeq = tonumber(nGrp.nextSeq) or 0
+      if nextExpectedSeq > 0 then
+        local relAdv = advancePastSkips(nextGroupHashKey, nextExpectedSeq)
+        if relAdv > nextExpectedSeq then
+          redis.call('HSET', nextGroupHashKey, 'nextSeq', tostring(relAdv))
+          nextExpectedSeq = relAdv
+        end
+      end
+      if nextExpectedSeq > 0 and nextJobOrderingSeq > nextExpectedSeq then
+        redis.call('XACK', streamKey, group, nextEntryId)
+        redis.call('XDEL', streamKey, nextEntryId)
+        redis.call('ZADD', nextWaitListKey, nextJobOrderingSeq, nextJobId)
+        redis.call('HSET', nextJobKey, 'state', 'group-waiting')
+        return {'NEXT_NONE', jobId}
+      end
+      nextReturning = (nextExpectedSeq > 0 and nextJobOrderingSeq < nextExpectedSeq)
+    end
+    -- Concurrency gate (skip for returning step-jobs)
+    if nextMaxConc > 0 and nextActive >= nextMaxConc and not nextReturning then
+      redis.call('XACK', streamKey, group, nextEntryId)
+      redis.call('XDEL', streamKey, nextEntryId)
+      redis.call('ZADD', nextWaitListKey, groupqScore(nextJobKey, nextJobId), nextJobId)
+      redis.call('HSET', nextJobKey, 'state', 'group-waiting')
+      return {'NEXT_NONE', jobId}
+    end
+    -- Token bucket gate (read-only)
+    local nextTbCapacity = tonumber(nGrp.tbCapacity) or 0
+    local nextTbBlocked = false
+    local nextTbDelay = 0
+    local nextTbTokens = 0
+    local nextJobCostVal = 0
+    if nextTbCapacity > 0 then
+      nextTbTokens = tbRefill(nextGroupHashKey, nGrp, tonumber(timestamp))
+      nextJobCostVal = tonumber(redis.call('HGET', nextJobKey, 'cost')) or 1000
+      -- DLQ guard: cost > capacity
+      if nextJobCostVal > nextTbCapacity then
+        local nextProcessedOn = tonumber(redis.call('HGET', nextJobKey, 'processedOn')) or tonumber(timestamp)
+        redis.call('XACK', streamKey, group, nextEntryId)
+        redis.call('XDEL', streamKey, nextEntryId)
+        redis.call('ZADD', prefix .. 'failed', tonumber(timestamp), nextJobId)
+        redis.call('HSET', nextJobKey,
+          'state', 'failed',
+          'failedReason', 'cost exceeds token bucket capacity',
+          'finishedOn', tostring(timestamp))
+        if skipEvents ~= '1' then emitEvent(prefix .. 'events', 'failed', nextJobId, {'failedReason', 'cost exceeds token bucket capacity'}) end
+        if skipMetrics ~= '1' then recordMetrics(metricsKey, tonumber(timestamp), tonumber(timestamp) - nextProcessedOn) end
+        return {'NEXT_NONE', jobId}
+      end
+      if nextTbTokens < nextJobCostVal then
+        nextTbBlocked = true
+        local nextTbRefillRateVal = math.max(tonumber(nGrp.tbRefillRate) or 0, 1)
+        nextTbDelay = math.ceil((nextJobCostVal - nextTbTokens) * 1000 / nextTbRefillRateVal)
+      end
+    end
+    -- Sliding window gate (read-only)
+    local nextRateMax = tonumber(nGrp.rateMax) or 0
+    local nextRlBlocked = false
+    local nextRlDelay = 0
+    if nextRateMax > 0 then
+      local nextRateDuration = tonumber(nGrp.rateDuration) or 0
+      local nextRateWindowStart = tonumber(nGrp.rateWindowStart) or 0
+      local nextRateCount = tonumber(nGrp.rateCount) or 0
+      if nextRateDuration > 0 and timestamp - nextRateWindowStart < nextRateDuration and nextRateCount >= nextRateMax then
+        nextRlBlocked = true
+        nextRlDelay = (nextRateWindowStart + nextRateDuration) - timestamp
+      end
+    end
+    -- If ANY gate blocked: park + register
+    if nextTbBlocked or nextRlBlocked then
+      redis.call('XACK', streamKey, group, nextEntryId)
+      redis.call('XDEL', streamKey, nextEntryId)
+      local nextWaitListKey = prefix .. 'groupq:' .. nextGroupKey
+      redis.call('ZADD', nextWaitListKey, groupqScore(nextJobKey, nextJobId), nextJobId)
+      redis.call('HSET', nextJobKey, 'state', 'group-waiting')
+      local nextMaxDelay = math.max(nextTbDelay, nextRlDelay)
+      local rateLimitedKey = prefix .. 'ratelimited'
+      redis.call('ZADD', rateLimitedKey, tonumber(timestamp) + nextMaxDelay, nextGroupKey)
+      return {'NEXT_NONE', jobId}
+    end
+    -- All gates passed: mutate state
+    if nextTbCapacity > 0 then
+      redis.call('HINCRBY', nextGroupHashKey, 'tbTokens', -nextJobCostVal)
+    end
+    if nextRateMax > 0 then
+      local nextRateDuration = tonumber(nGrp.rateDuration) or 0
+      if nextRateDuration > 0 then
+        local nextRateWindowStart = tonumber(nGrp.rateWindowStart) or 0
+        if timestamp - nextRateWindowStart >= nextRateDuration then
+          redis.call('HSET', nextGroupHashKey, 'rateWindowStart', tostring(timestamp), 'rateCount', '1')
+        else
+          redis.call('HINCRBY', nextGroupHashKey, 'rateCount', 1)
+        end
+      end
+    end
+    if not nextReturning then
+      redis.call('HINCRBY', nextGroupHashKey, 'active', 1)
+    end
+    if nextJobOrderingSeq > 0 and not nextReturning then
+      redis.call('HSET', nextGroupHashKey, 'nextSeq', tostring(nextJobOrderingSeq + 1))
+      redis.call('ZREM', nextWaitListKey, nextJobId)
+    end
+  end
+  redis.call('HSET', nextJobKey, 'state', 'active', 'processedOn', tostring(timestamp), 'lastActive', tostring(timestamp))
+  local nextHash = redis.call('HGETALL', nextJobKey)
+  local out = {'NEXT_HASH', jobId, nextJobId, nextEntryId}
+  for i = 1, #nextHash do
+    out[#out + 1] = nextHash[i]
+  end
+  return out
 end)
 
--- glidemq_fail(KEYS, ARGS)
--- Handles job failure: retry with backoff or move to failed set.
--- KEYS[1] = stream key
--- KEYS[2] = failed zset key
--- KEYS[3] = scheduled zset key
--- KEYS[4] = events stream key
--- KEYS[5] = job hash key (job:{id})
--- ARGS[1] = job ID
--- ARGS[2] = stream entry ID
--- ARGS[3] = failed reason (string)
--- ARGS[4] = timestamp (ms)
--- ARGS[5] = max attempts (0 = no retry)
--- ARGS[6] = backoff delay (ms, 0 = no backoff)
--- ARGS[7] = consumer group name
--- ARGS[8] = removeMode ('0'|'true'|'count'|'age_count')
--- ARGS[9] = removeCount (number)
--- ARGS[10] = removeAge (seconds)
--- Returns: "failed" | "retrying"
 redis.register_function('glidemq_fail', function(keys, args)
   local streamKey = keys[1]
   local failedKey = keys[2]
   local scheduledKey = keys[3]
   local eventsKey = keys[4]
   local jobKey = keys[5]
-
+  local metricsKey = keys[6]
   local jobId = args[1]
+  assert(string.sub(jobKey, -(4 + #jobId)) == 'job:' .. jobId, 'unexpected key format: ' .. jobKey)
   local entryId = args[2]
   local failedReason = args[3]
   local timestamp = tonumber(args[4])
@@ -411,36 +1362,51 @@ redis.register_function('glidemq_fail', function(keys, args)
   local removeMode = args[8] or '0'
   local removeCount = tonumber(args[9]) or 0
   local removeAge = tonumber(args[10]) or 0
-
-  -- Acknowledge the stream entry
-  redis.call('XACK', streamKey, group, entryId)
-  redis.call('XDEL', streamKey, entryId)
-
-  -- Increment attempt count
-  local attemptsMade = redis.call('HINCRBY', jobKey, 'attemptsMade', 1)
-
+  local broadcastMode = args[11] or '0'
+  local processedOn = tonumber(redis.call('HGET', jobKey, 'processedOn')) or timestamp
+  if entryId ~= '' then redis.call('XACK', streamKey, group, entryId) end
+  if entryId ~= '' and broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
+  local attemptsMade
+  if broadcastMode == '1' then
+    local subKey = jobKey .. ':sub:' .. group
+    attemptsMade = redis.call('HINCRBY', subKey, 'a', 1)
+    redis.call('EXPIRE', subKey, 86400)
+  else
+    attemptsMade = redis.call('HINCRBY', jobKey, 'attemptsMade', 1)
+  end
   if maxAttempts > 0 and attemptsMade < maxAttempts then
-    -- Retry: schedule with backoff
+    -- Advance fallback chain if the job has fallbacks configured
+    local optsRaw = redis.call('HGET', jobKey, 'opts')
+    if optsRaw and string.find(optsRaw, '"fallbacks"') then
+      local fbIdx = tonumber(redis.call('HGET', jobKey, 'fallbackIndex') or '0') or 0
+      redis.call('HSET', jobKey, 'fallbackIndex', tostring(fbIdx + 1))
+    end
     local retryAt = timestamp + backoffDelay
     local priority = tonumber(redis.call('HGET', jobKey, 'priority')) or 0
     local score = priority * PRIORITY_SHIFT + retryAt
-
-    redis.call('ZADD', scheduledKey, score, jobId)
+    local member = jobId
+    if broadcastMode == '1' then
+      member = jobId .. '||' .. group
+    end
+    redis.call('ZADD', scheduledKey, score, member)
     redis.call('HSET', jobKey,
       'state', 'delayed',
       'failedReason', failedReason,
       'processedOn', tostring(timestamp)
     )
-
+    -- Only release group slot if not an ordering-key job (ordering jobs hold the slot through retries)
+    local failOrdSeq = tonumber(redis.call('HGET', jobKey, 'orderingSeq')) or 0
+    if failOrdSeq <= 0 then
+      releaseGroupSlotAndPromote(jobKey, jobId, timestamp)
+    end
     emitEvent(eventsKey, 'retrying', jobId, {
       'failedReason', failedReason,
       'attemptsMade', tostring(attemptsMade),
       'delay', tostring(backoffDelay)
     })
-
+    if entryId == '' then decrListActive(string.sub(jobKey, 1, #jobKey - #('job:' .. jobId)) .. 'list-active') end
     return 'retrying'
   else
-    -- No more retries: move to failed set
     redis.call('ZADD', failedKey, timestamp, jobId)
     redis.call('HSET', jobKey,
       'state', 'failed',
@@ -449,90 +1415,66 @@ redis.register_function('glidemq_fail', function(keys, args)
       'processedOn', tostring(timestamp)
     )
     markOrderingDone(jobKey, jobId)
-
+    releaseGroupSlotAndPromote(jobKey, jobId, timestamp)
     emitEvent(eventsKey, 'failed', jobId, {'failedReason', failedReason})
-
-    -- Retention cleanup
+    recordMetrics(metricsKey, timestamp, timestamp - processedOn)
     local prefix = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
-    if removeMode == 'true' then
-      redis.call('ZREM', failedKey, jobId)
-      redis.call('UNLINK', jobKey)
-    elseif removeMode == 'count' and removeCount > 0 then
-      local total = redis.call('ZCARD', failedKey)
-      if total > removeCount then
-        local excess = redis.call('ZRANGE', failedKey, 0, total - removeCount - 1)
-        for i = 1, #excess do
-          local oldId = excess[i]
-          redis.call('UNLINK', prefix .. 'job:' .. oldId)
-          redis.call('ZREM', failedKey, oldId)
-        end
-      end
-    elseif removeMode == 'age_count' then
-      if removeAge > 0 then
-        local cutoff = timestamp - (removeAge * 1000)
-        local old = redis.call('ZRANGEBYSCORE', failedKey, '0', tostring(cutoff))
-        for i = 1, #old do
-          local oldId = old[i]
-          redis.call('UNLINK', prefix .. 'job:' .. oldId)
-          redis.call('ZREM', failedKey, oldId)
-        end
-      end
-      if removeCount > 0 then
+    -- In broadcast mode, skip job hash deletion: the job must persist for all subscriptions
+    if broadcastMode ~= '1' then
+      if removeMode == 'true' then
+        redis.call('ZREM', failedKey, jobId)
+        redis.call('UNLINK', jobKey)
+      elseif removeMode == 'count' and removeCount > 0 then
         local total = redis.call('ZCARD', failedKey)
         if total > removeCount then
-          local excess = redis.call('ZRANGE', failedKey, 0, total - removeCount - 1)
-          for i = 1, #excess do
-            local oldId = excess[i]
-            redis.call('UNLINK', prefix .. 'job:' .. oldId)
-            redis.call('ZREM', failedKey, oldId)
+          local excess = redis.call('ZRANGE', failedKey, 0, math.min(total - removeCount, 1000) - 1)
+          if #excess > 0 then removeExcessJobs(failedKey, prefix, excess) end
+        end
+      elseif removeMode == 'age_count' then
+        if removeAge > 0 then
+          local cutoff = timestamp - (removeAge * 1000)
+          local old = redis.call('ZRANGEBYSCORE', failedKey, '0', string.format('%.0f', cutoff), 'LIMIT', 0, 1000)
+          if #old > 0 then removeExcessJobs(failedKey, prefix, old) end
+        end
+        if removeCount > 0 then
+          local total = redis.call('ZCARD', failedKey)
+          if total > removeCount then
+            local excess = redis.call('ZRANGE', failedKey, 0, math.min(total - removeCount, 1000) - 1)
+            if #excess > 0 then removeExcessJobs(failedKey, prefix, excess) end
           end
         end
       end
     end
-
+    if entryId == '' then decrListActive(prefix .. 'list-active') end
     return 'failed'
   end
 end)
 
--- glidemq_reclaimStalled(KEYS, ARGS)
--- Uses XAUTOCLAIM to reclaim stalled jobs from inactive consumers.
--- KEYS[1] = stream key
--- KEYS[2] = events stream key
--- ARGS[1] = consumer group name
--- ARGS[2] = claiming consumer name
--- ARGS[3] = min idle time (ms)
--- ARGS[4] = max stalled count (move to failed after this many reclaims)
--- ARGS[5] = timestamp (ms)
--- ARGS[6] = failed zset key
--- Returns: number of reclaimed jobs
 redis.register_function('glidemq_reclaimStalled', function(keys, args)
   local streamKey = keys[1]
   local eventsKey = keys[2]
-
   local group = args[1]
   local consumer = args[2]
   local minIdleMs = tonumber(args[3])
   local maxStalledCount = tonumber(args[4]) or 1
   local timestamp = tonumber(args[5])
   local failedKey = args[6]
-
-  -- XAUTOCLAIM returns: [nextStartId, [[entryId, fields...], ...], [deletedIds...]]
+  local broadcastMode = args[7] or '0'
+  -- Worker-level lockDuration. Used as the per-entry stall threshold when the
+  -- job has no opts.lockDuration of its own. Falls back to minIdleMs (the
+  -- stalledInterval cadence) when neither is set, matching old behavior. (#213)
+  local workerLockDuration = tonumber(args[8]) or 0
   local result = redis.call('XAUTOCLAIM', streamKey, group, consumer, minIdleMs, '0-0')
-
   local entries = result[2]
   if not entries or #entries == 0 then
     return 0
   end
-
-  local prefix = string.sub(streamKey, 1, #streamKey - 6) -- strip "stream"
+  local prefix = string.sub(streamKey, 1, #streamKey - 6)
   local count = 0
-
   for i = 1, #entries do
     local entry = entries[i]
     local entryId = entry[1]
     local fields = entry[2]
-
-    -- Extract jobId from stream entry fields
     local jobId = nil
     if type(fields) == 'table' then
       for j = 1, #fields, 2 do
@@ -542,128 +1484,169 @@ redis.register_function('glidemq_reclaimStalled', function(keys, args)
         end
       end
     end
-
     if jobId then
       local jobKey = prefix .. 'job:' .. jobId
-
-      -- Check if the job has a recent heartbeat (lastActive field).
-      -- If (now - lastActive) < minIdleMs, the job is still being processed
-      -- and should not be reclaimed.
-      local lastActive = tonumber(redis.call('HGET', jobKey, 'lastActive'))
-      if lastActive and (timestamp - lastActive) < minIdleMs then
-        -- Job has a recent heartbeat - skip reclaiming.
-        -- We need to reset the idle time on the PEL entry by re-claiming
-        -- to the same consumer so it doesn't get picked up again next cycle.
-        -- XCLAIM already happened via XAUTOCLAIM, so just skip the stall logic.
+      if checkExpired(jobKey, jobId, prefix, timestamp) then
+        if entryId ~= '' then redis.call('XACK', streamKey, group, entryId) end
+        if entryId ~= '' and broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
         count = count + 1
       else
-
-      -- Track stall count on the job hash
-      local stalledCount = redis.call('HINCRBY', jobKey, 'stalledCount', 1)
-
-      if stalledCount > maxStalledCount then
-        -- Exceeded max stalled count: move to failed
-        redis.call('XACK', streamKey, group, entryId)
-        redis.call('XDEL', streamKey, entryId)
-        redis.call('ZADD', failedKey, timestamp, jobId)
-        redis.call('HSET', jobKey,
-          'state', 'failed',
-          'failedReason', 'job stalled more than maxStalledCount',
-          'finishedOn', tostring(timestamp)
-        )
-        markOrderingDone(jobKey, jobId)
-        emitEvent(eventsKey, 'failed', jobId, {
-          'failedReason', 'job stalled more than maxStalledCount'
-        })
+      local vals = redis.call('HMGET', jobKey, 'lastActive', 'opts')
+      local lastActive = tonumber(vals[1])
+      local jobLockDuration = extractLockDurationFromOpts(vals[2])
+      local effectiveIdle
+      if jobLockDuration > 0 then
+        effectiveIdle = jobLockDuration
+      elseif workerLockDuration > 0 then
+        effectiveIdle = workerLockDuration
+      else
+        effectiveIdle = minIdleMs
+      end
+      if lastActive and (timestamp - lastActive) < effectiveIdle then
+        count = count + 1
+      else
+      local failed = applyStalledLogic(jobKey, jobId, prefix, eventsKey, failedKey, maxStalledCount, timestamp)
+      if failed then
+        if entryId ~= '' then redis.call('XACK', streamKey, group, entryId) end
+        if entryId ~= '' and broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
       elseif broadcastMode == '1' then
         -- Broadcast mode keeps the original entry visible to all subscribers,
-        -- so we can't redispatch via XADD. Leave state=active and re-emit the
-        -- stalled event; if it stalls again next cycle, the > threshold branch
-        -- will mark it failed.
+        -- so we can't redispatch via XADD. Leave state active for the next
+        -- reclaim cycle; if it stalls again, applyStalledLogic will fail it.
         redis.call('HSET', jobKey, 'state', 'active')
-        emitEvent(eventsKey, 'stalled', jobId, nil)
       else
-        -- Under threshold: ACK+DEL the original PEL entry (which is now in
-        -- the scheduler's consumer after XAUTOCLAIM) and re-queue a fresh
-        -- stream entry so a healthy worker picks the job up. processedOn
-        -- and lastActive are cleared so the new run records its own metrics.
-        redis.call('XACK', streamKey, group, entryId)
-        redis.call('XDEL', streamKey, entryId)
+        -- Under maxStalledCount threshold: redispatch the job to the stream
+        -- so a healthy worker picks it up. The original PEL entry now belongs
+        -- to the scheduler's consumer (XAUTOCLAIM moved it there) and would
+        -- otherwise sit unprocessed forever; ACK+DEL it and re-queue a fresh
+        -- stream entry. processedOn is cleared so metrics record the new run.
+        if entryId ~= '' then redis.call('XACK', streamKey, group, entryId) end
+        if entryId ~= '' then redis.call('XDEL', streamKey, entryId) end
         local jobName = redis.call('HGET', jobKey, 'name') or ''
         redis.call('HSET', jobKey, 'state', 'waiting')
         redis.call('HDEL', jobKey, 'processedOn', 'lastActive')
-        redis.call('XADD', streamKey, '*', 'jobId', jobId, 'name', jobName)
-        emitEvent(eventsKey, 'stalled', jobId, nil)
+        xaddJob(streamKey, jobId, jobName)
       end
-
       count = count + 1
-      end -- close lastActive else
+      end
+      end
     end
   end
-
   return count
 end)
 
--- glidemq_pause(KEYS, ARGS)
--- Pauses a queue by setting paused=1 in the meta hash.
--- KEYS[1] = meta hash key
--- KEYS[2] = events stream key
--- ARGS: (none)
--- Returns: 1
+-- Reclaim stalled list-sourced jobs (LIFO/priority) that are invisible to XAUTOCLAIM.
+-- Uses bounded SCAN to find active list jobs with stale lastActive, then applies stall logic.
+-- KEYS: [streamKey, eventsKey]
+-- ARGS: [minIdleMs, maxStalledCount, timestamp, failedKey]
+redis.register_function('glidemq_reclaimStalledListJobs', function(keys, args)
+  local streamKey = keys[1]
+  local eventsKey = keys[2]
+  local minIdleMs = tonumber(args[1])
+  local maxStalledCount = tonumber(args[2]) or 1
+  local timestamp = tonumber(args[3])
+  if not minIdleMs or not timestamp then return 0 end
+  local failedKey = args[4]
+  -- Worker-level lockDuration; per-entry threshold when opts.lockDuration unset. (#213)
+  local workerLockDuration = tonumber(args[5]) or 0
+  local prefix = string.sub(streamKey, 1, #streamKey - 6)
+  local listActiveKey = prefix .. 'list-active'
+  local currentActive = tonumber(redis.call('GET', listActiveKey)) or 0
+  if currentActive <= 0 then return 0 end
+  local pattern = prefix .. 'job:*'
+  local cursor = '0'
+  local maxIter = 1000
+  local iter = 0
+  local count = 0
+  repeat
+    iter = iter + 1
+    local sr = redis.call('SCAN', cursor, 'MATCH', pattern, 'COUNT', 500)
+    cursor = sr[1]
+    local scannedKeys = sr[2]
+    for i = 1, #scannedKeys do
+      local jk = scannedKeys[i]
+      local state = redis.call('HGET', jk, 'state')
+      if state == 'active' then
+        local vals = redis.call('HMGET', jk, 'lastActive', 'lifo', 'priority', 'opts')
+        local lastActive = tonumber(vals[1])
+        local isListSourced = vals[2] == '1' or (tonumber(vals[3]) or 0) > 0
+        local jobLockDuration = extractLockDurationFromOpts(vals[4])
+        local effectiveIdle
+        if jobLockDuration > 0 then
+          effectiveIdle = jobLockDuration
+        elseif workerLockDuration > 0 then
+          effectiveIdle = workerLockDuration
+        else
+          effectiveIdle = minIdleMs
+        end
+        if isListSourced and lastActive and (timestamp - lastActive) >= effectiveIdle then
+          -- Claim this stalled check by refreshing lastActive. This preserves
+          -- XAUTOCLAIM-like de-duplication semantics across worker schedulers
+          -- so the same stale list job is counted at most once per interval.
+          redis.call('HSET', jk, 'lastActive', tostring(timestamp))
+          local jobId = string.sub(jk, #prefix + 5)
+          local shouldDecr = false
+          if checkExpired(jk, jobId, prefix, timestamp) then
+            shouldDecr = true
+          else
+            if applyStalledLogic(jk, jobId, prefix, eventsKey, failedKey, maxStalledCount, timestamp) then
+              shouldDecr = true
+            else
+              -- Under threshold: re-queue the job onto its source list so a
+              -- healthy worker picks it up. LIFO -> RPUSH lifo, priority ->
+              -- LPUSH priority (RPOP returns highest priority first). Decrement
+              -- the list-active counter because the job is no longer active.
+              local isLifo = vals[2] == '1'
+              local pri = tonumber(vals[3]) or 0
+              redis.call('HSET', jk, 'state', 'waiting')
+              redis.call('HDEL', jk, 'processedOn', 'lastActive')
+              if isLifo then
+                redis.call('RPUSH', prefix .. 'lifo', jobId)
+              elseif pri > 0 then
+                redis.call('LPUSH', prefix .. 'priority', jobId)
+              else
+                -- Stream-sourced fallback (shouldn't normally hit this branch
+                -- because XAUTOCLAIM handles stream entries, but redispatch
+                -- safely if someone misclassifies).
+                redis.call('XADD', streamKey, '*', 'jobId', jobId, 'name',
+                  redis.call('HGET', jk, 'name') or '')
+              end
+              shouldDecr = true
+            end
+          end
+          if shouldDecr then
+            decrListActive(listActiveKey)
+          end
+          count = count + 1
+        end
+      end
+    end
+  until cursor == '0' or iter >= maxIter
+  return count
+end)
+
 redis.register_function('glidemq_pause', function(keys, args)
   local metaKey = keys[1]
   local eventsKey = keys[2]
-
   redis.call('HSET', metaKey, 'paused', '1')
   emitEvent(eventsKey, 'paused', '0', nil)
-
   return 1
 end)
 
--- glidemq_resume(KEYS, ARGS)
--- Resumes a queue by setting paused=0 in the meta hash.
--- KEYS[1] = meta hash key
--- KEYS[2] = events stream key
--- ARGS: (none)
--- Returns: 1
 redis.register_function('glidemq_resume', function(keys, args)
   local metaKey = keys[1]
   local eventsKey = keys[2]
-
   redis.call('HSET', metaKey, 'paused', '0')
   emitEvent(eventsKey, 'resumed', '0', nil)
-
   return 1
 end)
 
--- glidemq_dedup(KEYS, ARGS)
--- Deduplication: check dedup hash and either skip or add a new job.
--- KEYS[1] = dedup hash key
--- KEYS[2] = id counter key
--- KEYS[3] = stream key
--- KEYS[4] = scheduled zset key
--- KEYS[5] = events stream key
--- ARGS[1] = dedup_id
--- ARGS[2] = ttl_ms (0 = no TTL, used by throttle mode)
--- ARGS[3] = mode ("simple" | "throttle" | "debounce")
--- ARGS[4] = job name
--- ARGS[5] = job data (JSON string)
--- ARGS[6] = job opts (JSON string)
--- ARGS[7] = timestamp (ms)
--- ARGS[8] = delay (ms, 0 = no delay)
--- ARGS[9] = priority (0 = no priority)
--- ARGS[10] = parentId (empty string = no parent)
--- ARGS[11] = maxAttempts (0 = no retry)
--- ARGS[12] = orderingKey (empty string = no per-key ordering)
--- ARGS[20] = customJobId (empty string = auto-generate via INCR)
--- Returns: "skipped" if deduplicated, "duplicate" if custom ID exists, otherwise the new job ID (string)
 redis.register_function('glidemq_dedup', function(keys, args)
   local dedupKey = keys[1]
   local idKey = keys[2]
   local streamKey = keys[3]
   local scheduledKey = keys[4]
   local eventsKey = keys[5]
-
   local dedupId = args[1]
   local ttlMs = tonumber(args[2]) or 0
   local mode = args[3]
@@ -676,15 +1659,20 @@ redis.register_function('glidemq_dedup', function(keys, args)
   local parentId = args[10] or ''
   local maxAttempts = tonumber(args[11]) or 0
   local orderingKey = args[12] or ''
+  local groupConcurrency = tonumber(args[13]) or 0
+  local lifo = tonumber(args[21]) or 0
+  local groupRateMax = tonumber(args[14]) or 0
+  local groupRateDuration = tonumber(args[15]) or 0
+  local tbCapacity = tonumber(args[16]) or 0
+  local tbRefillRate = tonumber(args[17]) or 0
+  local jobCost = tonumber(args[18]) or 0
+  local ttl = tonumber(args[19]) or 0
   local customJobId = args[20] or ''
-
-  local prefix = string.sub(idKey, 1, #idKey - 2) -- strip ":id"
-
-  -- Check existing dedup entry: value is "jobId:timestamp"
+  local parentQueue = args[22] or ''
+  local skipEvents = args[23] or '0'
+  local prefix = string.sub(idKey, 1, #idKey - 2)
   local existing = redis.call('HGET', dedupKey, dedupId)
-
   if mode == 'simple' then
-    -- Simple: skip if dedup_id exists and job is not completed/failed
     if existing then
       local sep = string.find(existing, ':')
       if sep then
@@ -697,7 +1685,6 @@ redis.register_function('glidemq_dedup', function(keys, args)
       end
     end
   elseif mode == 'throttle' then
-    -- Throttle: skip if (now - stored_timestamp) < ttl
     if existing and ttlMs > 0 then
       local sep = string.find(existing, ':')
       if sep then
@@ -708,7 +1695,6 @@ redis.register_function('glidemq_dedup', function(keys, args)
       end
     end
   elseif mode == 'debounce' then
-    -- Debounce: if exists and job is delayed, remove old job and replace with new one
     if existing then
       local sep = string.find(existing, ':')
       if sep then
@@ -716,20 +1702,22 @@ redis.register_function('glidemq_dedup', function(keys, args)
         local jobKey = prefix .. 'job:' .. existingJobId
         local state = redis.call('HGET', jobKey, 'state')
         if state == 'delayed' or state == 'prioritized' then
-          -- Remove old job from scheduled ZSet and delete hash
+          local delGroupKey = redis.call('HGET', jobKey, 'groupKey')
+          local delOrderingSeq = tonumber(redis.call('HGET', jobKey, 'orderingSeq')) or 0
           redis.call('ZREM', scheduledKey, existingJobId)
           markOrderingDone(jobKey, existingJobId)
+          if delGroupKey and delGroupKey ~= '' and delOrderingSeq > 0 then
+            local groupHashKey = prefix .. 'group:' .. delGroupKey
+            redis.call('HSET', groupHashKey, 'skip:' .. tostring(delOrderingSeq), '1')
+          end
           redis.call('UNLINK', jobKey)
-          emitEvent(eventsKey, 'removed', existingJobId, nil)
+          if skipEvents ~= '1' then emitEvent(eventsKey, 'removed', existingJobId, nil) end
         elseif state and state ~= 'completed' and state ~= 'failed' then
-          -- Job is waiting or active - skip (can't debounce non-delayed)
           return 'skipped'
         end
       end
     end
   end
-
-  -- Add the new job (same logic as glidemq_addJob)
   local jobIdStr
   local jobKey
   if customJobId ~= '' then
@@ -738,25 +1726,74 @@ redis.register_function('glidemq_dedup', function(keys, args)
       return 'duplicate'
     end
     jobIdStr = customJobId
+    advanceIdCounter(idKey, customJobId)
   else
     local jobId = redis.call('INCR', idKey)
     jobIdStr = tostring(jobId)
     jobKey = prefix .. 'job:' .. jobIdStr
-    local retries = 0
-    while redis.call('EXISTS', jobKey) == 1 do
-      retries = retries + 1
-      if retries >= 1000 then return 'ERR:ID_EXHAUSTED' end
-      jobId = redis.call('INCR', idKey)
-      jobIdStr = tostring(jobId)
-      jobKey = prefix .. 'job:' .. jobIdStr
-    end
   end
+  local useGroupConcurrency = (orderingKey ~= '')
   local orderingSeq = 0
   if orderingKey ~= '' then
     local orderingMetaKey = prefix .. 'ordering'
     orderingSeq = redis.call('HINCRBY', orderingMetaKey, orderingKey, 1)
   end
-
+  if useGroupConcurrency then
+    if groupConcurrency < 1 then groupConcurrency = 1 end
+    local groupHashKey = prefix .. 'group:' .. orderingKey
+    local curMax = tonumber(redis.call('HGET', groupHashKey, 'maxConcurrency')) or 0
+    if curMax ~= groupConcurrency then
+      redis.call('HSET', groupHashKey, 'maxConcurrency', tostring(groupConcurrency))
+    end
+    -- Initialize nextSeq for ordered promotion
+    if orderingSeq > 0 and redis.call('HEXISTS', groupHashKey, 'nextSeq') == 0 then
+      redis.call('HSET', groupHashKey, 'nextSeq', '1')
+    end
+    if groupRateMax > 0 then
+      local curRateMax = tonumber(redis.call('HGET', groupHashKey, 'rateMax')) or 0
+      if curRateMax ~= groupRateMax then
+        redis.call('HSET', groupHashKey, 'rateMax', tostring(groupRateMax))
+      end
+      local curRateDuration = tonumber(redis.call('HGET', groupHashKey, 'rateDuration')) or 0
+      if curRateDuration ~= groupRateDuration then
+        redis.call('HSET', groupHashKey, 'rateDuration', tostring(groupRateDuration))
+      end
+    else
+      local oldRateMax = tonumber(redis.call('HGET', groupHashKey, 'rateMax')) or 0
+      if oldRateMax > 0 then
+        redis.call('HDEL', groupHashKey, 'rateMax', 'rateDuration', 'rateWindowStart', 'rateCount')
+      end
+    end
+    -- Upsert token bucket fields on group hash
+    if tbCapacity > 0 then
+      local curTbCap = tonumber(redis.call('HGET', groupHashKey, 'tbCapacity')) or 0
+      if curTbCap ~= tbCapacity then
+        redis.call('HSET', groupHashKey, 'tbCapacity', tostring(tbCapacity))
+      end
+      local curTbRate = tonumber(redis.call('HGET', groupHashKey, 'tbRefillRate')) or 0
+      if curTbRate ~= tbRefillRate then
+        redis.call('HSET', groupHashKey, 'tbRefillRate', tostring(tbRefillRate))
+      end
+      -- Initialize tokens on first setup
+      if curTbCap == 0 then
+        redis.call('HSET', groupHashKey,
+          'tbTokens', tostring(tbCapacity),
+          'tbLastRefill', tostring(timestamp),
+          'tbRefillRemainder', '0')
+      end
+      -- Validate cost (explicit or default 1000 millitokens) against capacity
+      local effectiveCost = (jobCost > 0) and jobCost or 1000
+      if effectiveCost > tbCapacity then
+        return 'ERR:COST_EXCEEDS_CAPACITY'
+      end
+    else
+      -- Clear stale tb fields
+      local oldTbCap = tonumber(redis.call('HGET', groupHashKey, 'tbCapacity')) or 0
+      if oldTbCap > 0 then
+        redis.call('HDEL', groupHashKey, 'tbCapacity', 'tbRefillRate', 'tbTokens', 'tbLastRefill', 'tbRefillRemainder')
+      end
+    end
+  end
   local hashFields = {
     'id', jobIdStr,
     'name', jobName,
@@ -768,19 +1805,34 @@ redis.register_function('glidemq_dedup', function(keys, args)
     'priority', tostring(priority),
     'maxAttempts', tostring(maxAttempts)
   }
-
-  if orderingKey ~= '' then
-    hashFields[#hashFields + 1] = 'orderingKey'
+  if useGroupConcurrency then
+    hashFields[#hashFields + 1] = 'groupKey'
     hashFields[#hashFields + 1] = orderingKey
-    hashFields[#hashFields + 1] = 'orderingSeq'
-    hashFields[#hashFields + 1] = tostring(orderingSeq)
+    if orderingSeq > 0 then
+      hashFields[#hashFields + 1] = 'orderingSeq'
+      hashFields[#hashFields + 1] = tostring(orderingSeq)
+    end
   end
-
+  if jobCost > 0 then
+    hashFields[#hashFields + 1] = 'cost'
+    hashFields[#hashFields + 1] = tostring(jobCost)
+  end
+  if ttl > 0 then
+    hashFields[#hashFields + 1] = 'expireAt'
+    hashFields[#hashFields + 1] = tostring(timestamp + ttl)
+  end
   if parentId ~= '' then
     hashFields[#hashFields + 1] = 'parentId'
     hashFields[#hashFields + 1] = parentId
+    if parentQueue ~= '' then
+      hashFields[#hashFields + 1] = 'parentQueue'
+      hashFields[#hashFields + 1] = parentQueue
+    end
   end
-
+  if lifo > 0 then
+    hashFields[#hashFields + 1] = 'lifo'
+    hashFields[#hashFields + 1] = '1'
+  end
   if delay > 0 or priority > 0 then
     hashFields[#hashFields + 1] = 'state'
     hashFields[#hashFields + 1] = delay > 0 and 'delayed' or 'prioritized'
@@ -788,160 +1840,461 @@ redis.register_function('glidemq_dedup', function(keys, args)
     hashFields[#hashFields + 1] = 'state'
     hashFields[#hashFields + 1] = 'waiting'
   end
-
   redis.call('HSET', jobKey, unpack(hashFields))
-
+  -- Register child in parent's deps set when parentDepsKey is provided (keys[6])
+  if parentId ~= '' and parentQueue ~= '' and #keys >= 6 then
+    local parentDepsKey = keys[6]
+    local queuePrefix = string.sub(prefix, 1, #prefix - 1)
+    local depsMember = queuePrefix .. ':' .. jobIdStr
+    redis.call('SADD', parentDepsKey, depsMember)
+  end
   if delay > 0 then
     local score = priority * PRIORITY_SHIFT + (timestamp + delay)
     redis.call('ZADD', scheduledKey, score, jobIdStr)
   elseif priority > 0 then
     local score = priority * PRIORITY_SHIFT
     redis.call('ZADD', scheduledKey, score, jobIdStr)
+  elseif lifo > 0 then
+    local lifoKey = prefix .. 'lifo'
+    redis.call('RPUSH', lifoKey, jobIdStr)
   else
-    redis.call('XADD', streamKey, '*', 'jobId', jobIdStr)
+    xaddJob(streamKey, jobIdStr, jobName)
   end
-
-  -- Store dedup entry: field=dedup_id, value=jobId:timestamp
   redis.call('HSET', dedupKey, dedupId, jobIdStr .. ':' .. tostring(timestamp))
-
-  emitEvent(eventsKey, 'added', jobIdStr, {'name', jobName})
-
+  if skipEvents ~= '1' then emitEvent(eventsKey, 'added', jobIdStr, {'name', jobName}) end
   return jobIdStr
 end)
 
--- glidemq_rateLimit(KEYS, ARGS)
--- Checks and enforces a sliding-window rate limit.
--- KEYS[1] = rate hash key (glide:{q}:rate)
--- KEYS[2] = meta key (glide:{q}:meta)
--- ARGS[1] = max per window
--- ARGS[2] = window duration (ms)
--- ARGS[3] = current timestamp (ms)
--- Returns: 0 if allowed, positive number = ms to wait before retrying
 redis.register_function('glidemq_rateLimit', function(keys, args)
   local rateKey = keys[1]
   local metaKey = keys[2]
-
   local maxPerWindow = tonumber(args[1])
   local windowDuration = tonumber(args[2])
   local now = tonumber(args[3])
-
+  -- Fallback: read rate limit config from meta if not provided inline
+  if maxPerWindow <= 0 then
+    maxPerWindow = tonumber(redis.call('HGET', metaKey, 'rateLimitMax')) or 0
+    windowDuration = tonumber(redis.call('HGET', metaKey, 'rateLimitDuration')) or 0
+    if maxPerWindow <= 0 then return 0 end
+  end
   local windowStart = tonumber(redis.call('HGET', rateKey, 'windowStart')) or 0
   local count = tonumber(redis.call('HGET', rateKey, 'count')) or 0
-
-  -- If the current window has expired, reset
   if now - windowStart >= windowDuration then
     redis.call('HSET', rateKey, 'windowStart', tostring(now), 'count', '1')
     return 0
   end
-
-  -- Window still active: check count
   if count >= maxPerWindow then
-    -- Rate limited: return ms until window resets
     local delayMs = windowDuration - (now - windowStart)
     return delayMs
   end
-
-  -- Allowed: increment count
   redis.call('HSET', rateKey, 'count', tostring(count + 1))
   return 0
 end)
 
--- glidemq_checkConcurrency(KEYS, ARGS)
--- Checks global concurrency limit.
--- KEYS[1] = meta hash key
--- KEYS[2] = stream key
--- ARGS[1] = consumer group name
--- Returns: -1 if no limit set, 0 if blocked, or positive remaining capacity
+redis.register_function('glidemq_promoteRateLimited', function(keys, args)
+  local rateLimitedKey = keys[1]
+  local streamKey = keys[2]
+  local now = tonumber(args[1])
+  -- Derive prefix from the server-validated key instead of caller-supplied arg
+  local prefix = string.sub(rateLimitedKey, 1, #rateLimitedKey - #'ratelimited')
+  local expired = redis.call('ZRANGEBYSCORE', rateLimitedKey, '0', string.format('%.0f', now), 'LIMIT', 0, 100)
+  if not expired or #expired == 0 then return 0 end
+  local promoted = 0
+  for i = 1, #expired do
+    local gk = expired[i]
+    redis.call('ZREM', rateLimitedKey, gk)
+    local groupHashKey = prefix .. 'group:' .. gk
+    local waitListKey = prefix .. 'groupq:' .. gk
+    -- Load all group fields in one call for rate limit + token bucket checks
+    local prGrpFields = redis.call('HGETALL', groupHashKey)
+    local prGrp = {}
+    for pf = 1, #prGrpFields, 2 do prGrp[prGrpFields[pf]] = prGrpFields[pf + 1] end
+    local rateMax = tonumber(prGrp.rateMax) or 0
+    local maxConc = tonumber(prGrp.maxConcurrency) or 0
+    local active = tonumber(prGrp.active) or 0
+    -- Token bucket pre-check: peek head job cost before promoting
+    local prTbCap = tonumber(prGrp.tbCapacity) or 0
+    local tbCheckPassed = true
+    if prTbCap > 0 then
+      local prTbTokens = tbRefill(groupHashKey, prGrp, now)
+      local headMembers = redis.call('ZRANGE', waitListKey, 0, 0)
+      local headJobId = headMembers[1]
+      if headJobId then
+        local headJobKey = prefix .. 'job:' .. headJobId
+        -- Tombstone guard
+        if redis.call('EXISTS', headJobKey) == 0 then
+          redis.call('ZREM', waitListKey, headJobId)
+          tbCheckPassed = false
+        end
+        if tbCheckPassed then
+          local headCost = tonumber(redis.call('HGET', headJobKey, 'cost')) or 1000
+          -- DLQ guard: cost > capacity
+          if headCost > prTbCap then
+            redis.call('ZREM', waitListKey, headJobId)
+            redis.call('ZADD', prefix .. 'failed', now, headJobId)
+            redis.call('HSET', headJobKey,
+              'state', 'failed',
+              'failedReason', 'cost exceeds token bucket capacity',
+              'finishedOn', tostring(now))
+            emitEvent(prefix .. 'events', 'failed', headJobId, {'failedReason', 'cost exceeds token bucket capacity'})
+            tbCheckPassed = false
+          end
+          if tbCheckPassed and prTbTokens < headCost then
+            -- Not enough tokens: re-register with calculated delay
+            local prTbRate = math.max(tonumber(prGrp.tbRefillRate) or 0, 1)
+            local prTbDelay = math.ceil((headCost - prTbTokens) * 1000 / prTbRate)
+            redis.call('ZADD', rateLimitedKey, now + prTbDelay, gk)
+            tbCheckPassed = false
+          end
+        end
+      end
+    end
+    if tbCheckPassed then
+      -- Promote up to min(rateMax, available concurrency) jobs.
+      -- Do NOT touch rateCount/rateWindowStart here - moveToActive handles
+      -- window reset and counting when the worker picks up the promoted jobs.
+      local canPromote = 1000
+      if rateMax > 0 then
+        canPromote = math.min(canPromote, rateMax)
+      end
+      if maxConc > 0 then
+        canPromote = math.min(canPromote, math.max(0, maxConc - active))
+      end
+      local prNextSeq = tonumber(prGrp.nextSeq) or 0
+      if prNextSeq > 0 then
+        local prAdv = advancePastSkips(groupHashKey, prNextSeq)
+        if prAdv > prNextSeq then
+          redis.call('HSET', groupHashKey, 'nextSeq', tostring(prAdv))
+          prNextSeq = prAdv
+        end
+      end
+      local groupPromoted = 0
+      local prIter = 0
+      local prMaxIter = canPromote + 20
+      while groupPromoted < canPromote and prIter < prMaxIter do
+        prIter = prIter + 1
+        local zpResult = redis.call('ZPOPMIN', waitListKey, 1)
+        local nextJobId = zpResult[1]
+        if not nextJobId then break end
+        local nextJobKey = prefix .. 'job:' .. nextJobId
+        local nextState = redis.call('HGET', nextJobKey, 'state')
+        if nextState ~= 'group-waiting' then
+          -- Stale: skip
+        elseif checkExpired(nextJobKey, nextJobId, prefix, now) then
+          -- Expired: skip
+        else
+          if prNextSeq > 0 then
+            local jobSeq = tonumber(redis.call('HGET', nextJobKey, 'orderingSeq')) or 0
+            if jobSeq > prNextSeq then
+              redis.call('ZADD', waitListKey, jobSeq, nextJobId)
+              break
+            end
+          end
+          xaddJob(streamKey, nextJobId, redis.call('HGET', nextJobKey, 'name'))
+          redis.call('HSET', nextJobKey, 'state', 'waiting')
+          promoted = promoted + 1
+          groupPromoted = groupPromoted + 1
+          if prNextSeq > 0 then prNextSeq = prNextSeq + 1 end
+        end
+      end
+    end
+  end
+  return promoted
+end)
+
 redis.register_function('glidemq_checkConcurrency', function(keys, args)
   local metaKey = keys[1]
   local streamKey = keys[2]
+  local listActiveKey = keys[3]
   local group = args[1]
-
   local gc = tonumber(redis.call('HGET', metaKey, 'globalConcurrency')) or 0
   if gc <= 0 then
     return -1
   end
-
-  local pending = redis.call('XPENDING', streamKey, group)
-  local pendingCount = tonumber(pending[1]) or 0
-  local remaining = gc - pendingCount
+  local ok_cc, pending = pcall(redis.call, 'XPENDING', streamKey, group)
+  local pendingCount = (ok_cc and pending and tonumber(pending[1])) or 0
+  local listActive = tonumber(redis.call('GET', listActiveKey)) or 0
+  local remaining = gc - pendingCount - listActive
   if remaining <= 0 then
     return 0
   end
   return remaining
 end)
 
--- glidemq_addFlow(KEYS, ARGS)
--- Atomically creates a parent job (waiting-children) and N child jobs.
--- KEYS[1] = parent id counter key (glide:{parentQueue}:id)
--- KEYS[2] = parent stream key
--- KEYS[3] = parent scheduled key
--- KEYS[4] = parent events key
--- Remaining KEYS are passed per child: 4 keys each (id, stream, scheduled, events)
--- ARGS layout:
---   ARGS[1] = parent job name
---   ARGS[2] = parent job data (JSON)
---   ARGS[3] = parent job opts (JSON)
---   ARGS[4] = timestamp (ms)
---   ARGS[5] = parent delay (ms)
---   ARGS[6] = parent priority
---   ARGS[7] = parent maxAttempts
---   ARGS[8] = number of children (N)
---   Then for each child (8 args each):
---     childName, childData, childOpts, childDelay, childPriority, childMaxAttempts, childQueuePrefix, childParentQueue
--- glidemq_moveToActive(KEYS, ARGS)
--- Single round trip: read job hash, check revoked, set state=active + processedOn + lastActive.
--- KEYS[1] = job hash key
--- ARGS[1] = timestamp
--- Returns: '' if not found, 'REVOKED' if revoked, cjson array of HGETALL fields otherwise
+-- Atomically check global concurrency capacity, then RPOP up to count jobs from a list and INCRBY list-active counter.
+-- Returns an array of popped jobIds (may be empty if at capacity or list is empty).
+-- KEYS: [metaKey, streamKey, listActiveKey, listKey]
+-- ARGS: [group, count]
+redis.register_function('glidemq_rpopAndReserve', function(keys, args)
+  local metaKey = keys[1]
+  local streamKey = keys[2]
+  local listActiveKey = keys[3]
+  local listKey = keys[4]
+  local group = args[1]
+  local requested = tonumber(args[2]) or 1
+  local maxPop = requested
+  local gc = tonumber(redis.call('HGET', metaKey, 'globalConcurrency')) or 0
+  if gc > 0 then
+    local ok_ra, pending = pcall(redis.call, 'XPENDING', streamKey, group)
+    local pendingCount = (ok_ra and pending and tonumber(pending[1])) or 0
+    local listActive = tonumber(redis.call('GET', listActiveKey)) or 0
+    local available = gc - pendingCount - listActive
+    if available <= 0 then
+      return {}
+    end
+    if available < maxPop then
+      maxPop = available
+    end
+  end
+  local results = {}
+  for i = 1, maxPop do
+    local jobId = redis.call('RPOP', listKey)
+    if not jobId then break end
+    results[#results + 1] = jobId
+  end
+  if #results > 0 then
+    redis.call('INCRBY', listActiveKey, #results)
+  end
+  return results
+end)
+
 redis.register_function('glidemq_moveToActive', function(keys, args)
   local jobKey = keys[1]
+  local streamKey = keys[2] or ''
   local timestamp = args[1]
-  local exists = redis.call('EXISTS', jobKey)
-  if exists == 0 then
+  local entryId = args[2] or ''
+  local group = args[3] or ''
+  local jobId = args[4] or ''
+  local broadcastMode = args[5] or '0'
+  local ts = tonumber(timestamp) or 0
+  local timestampStr = tostring(ts)
+  local fields = redis.call('HGETALL', jobKey)
+  if not fields or #fields == 0 then
     return ''
   end
-  local revoked = redis.call('HGET', jobKey, 'revoked')
+  local revoked = ''
+  local expireAt = 0
+  local curState = ''
+  local orderingKey = ''
+  local orderingSeq = ''
+  local groupKey = ''
+  local costVal = ''
+  local stateValueIndex = nil
+  local processedOnValueIndex = nil
+  local lastActiveValueIndex = nil
+  for f = 1, #fields, 2 do
+    local field = fields[f]
+    local value = fields[f + 1]
+    if field == 'revoked' then
+      revoked = value
+    elseif field == 'expireAt' then
+      expireAt = tonumber(value) or 0
+    elseif field == 'state' then
+      curState = value
+      stateValueIndex = f + 1
+    elseif field == 'orderingKey' then
+      orderingKey = value
+    elseif field == 'orderingSeq' then
+      orderingSeq = value
+    elseif field == 'groupKey' then
+      groupKey = value
+    elseif field == 'cost' then
+      costVal = value
+    elseif field == 'processedOn' then
+      processedOnValueIndex = f + 1
+    elseif field == 'lastActive' then
+      lastActiveValueIndex = f + 1
+    end
+  end
   if revoked == '1' then
     return 'REVOKED'
   end
-  redis.call('HSET', jobKey, 'state', 'active', 'processedOn', timestamp, 'lastActive', timestamp)
-  local fields = redis.call('HGETALL', jobKey)
-  return cjson.encode(fields)
+  local prefix = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
+  if expireAt > 0 and ts > expireAt then
+    expireJob(jobKey, jobId, prefix, ts, curState, orderingKey, orderingSeq, groupKey)
+    if streamKey ~= '' and entryId ~= '' and group ~= '' then
+      redis.call('XACK', streamKey, group, entryId)
+      if broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
+    end
+    return 'EXPIRED'
+  end
+  if groupKey and groupKey ~= '' then
+    local groupHashKey = prefix .. 'group:' .. groupKey
+    -- Load all group fields in one call
+    local grpFields = redis.call('HGETALL', groupHashKey)
+    local grp = {}
+    for f = 1, #grpFields, 2 do grp[grpFields[f]] = grpFields[f + 1] end
+    local maxConc = tonumber(grp.maxConcurrency) or 0
+    local active = tonumber(grp.active) or 0
+    local waitListKey = prefix .. 'groupq:' .. groupKey
+    -- Ordering gate: park future jobs (seq > nextSeq) in sorted groupq.
+    -- Returning step-jobs (seq <= nextSeq) are allowed through.
+    local jobOrderingSeq = tonumber(redis.call('HGET', jobKey, 'orderingSeq')) or 0
+    local isReturningStepJob = false
+    if jobOrderingSeq > 0 then
+      local nextSeq = tonumber(grp.nextSeq) or 0
+      if nextSeq > 0 then
+        local mAdv = advancePastSkips(groupHashKey, nextSeq)
+        if mAdv > nextSeq then
+          redis.call('HSET', groupHashKey, 'nextSeq', tostring(mAdv))
+          nextSeq = mAdv
+        end
+      end
+      if nextSeq > 0 and jobOrderingSeq > nextSeq then
+        if streamKey ~= '' and entryId ~= '' and group ~= '' then
+          redis.call('XACK', streamKey, group, entryId)
+          if broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
+        end
+        redis.call('ZADD', waitListKey, jobOrderingSeq, jobId)
+        redis.call('HSET', jobKey, 'state', 'group-waiting')
+        return 'GROUP_ORDERED'
+      end
+      isReturningStepJob = (nextSeq > 0 and jobOrderingSeq < nextSeq)
+    end
+    -- Concurrency gate (skip for returning step-jobs that already hold the slot)
+    if maxConc > 0 and active >= maxConc and not isReturningStepJob then
+      if streamKey ~= '' and entryId ~= '' and group ~= '' then
+        redis.call('XACK', streamKey, group, entryId)
+        if broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
+      end
+      redis.call('ZADD', waitListKey, groupqScore(jobKey, jobId), jobId)
+      redis.call('HSET', jobKey, 'state', 'group-waiting')
+      return 'GROUP_FULL'
+    end
+    -- Token bucket gate (read-only)
+    local tbCapacity = tonumber(grp.tbCapacity) or 0
+    local tbBlocked = false
+    local tbDelay = 0
+    local tbTokens = 0
+    local jobCostVal = 0
+    if tbCapacity > 0 then
+      tbTokens = tbRefill(groupHashKey, grp, ts)
+      jobCostVal = tonumber(costVal) or 1000
+      -- DLQ guard: cost > capacity
+      if jobCostVal > tbCapacity then
+        if streamKey ~= '' and entryId ~= '' and group ~= '' then
+          redis.call('XACK', streamKey, group, entryId)
+          if broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
+        end
+        redis.call('ZADD', prefix .. 'failed', ts, jobId)
+        redis.call('HSET', jobKey,
+          'state', 'failed',
+          'failedReason', 'cost exceeds token bucket capacity',
+          'finishedOn', timestampStr)
+        emitEvent(prefix .. 'events', 'failed', jobId, {'failedReason', 'cost exceeds token bucket capacity'})
+        return 'ERR:COST_EXCEEDS_CAPACITY'
+      end
+      if tbTokens < jobCostVal then
+        tbBlocked = true
+        local tbRefillRateVal = tonumber(grp.tbRefillRate) or 0
+        if tbRefillRateVal <= 0 then tbRefillRateVal = 1 end
+        tbDelay = math.ceil((jobCostVal - tbTokens) * 1000 / tbRefillRateVal)
+      end
+    end
+    -- Sliding window gate (read-only)
+    local rateMax = tonumber(grp.rateMax) or 0
+    local rlBlocked = false
+    local rlDelay = 0
+    if rateMax > 0 then
+      local rateDuration = tonumber(grp.rateDuration) or 0
+      local rateWindowStart = tonumber(grp.rateWindowStart) or 0
+      local rateCount = tonumber(grp.rateCount) or 0
+      local now = ts
+      if rateDuration > 0 and now - rateWindowStart < rateDuration and rateCount >= rateMax then
+        rlBlocked = true
+        rlDelay = (rateWindowStart + rateDuration) - now
+      end
+    end
+    -- If ANY gate blocked: park + register (skip for returning step-jobs)
+    if (tbBlocked or rlBlocked) and not isReturningStepJob then
+      if streamKey ~= '' and entryId ~= '' and group ~= '' then
+        redis.call('XACK', streamKey, group, entryId)
+        if broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
+      end
+      local waitListKey = prefix .. 'groupq:' .. groupKey
+      redis.call('ZADD', waitListKey, groupqScore(jobKey, jobId), jobId)
+      redis.call('HSET', jobKey, 'state', 'group-waiting')
+      local maxDelay = math.max(tbDelay, rlDelay)
+      local rateLimitedKey = prefix .. 'ratelimited'
+      redis.call('ZADD', rateLimitedKey, ts + maxDelay, groupKey)
+      if tbBlocked then return 'GROUP_TOKEN_LIMITED' end
+      return 'GROUP_RATE_LIMITED'
+    end
+    -- All gates passed: mutate state
+    if tbCapacity > 0 then
+      redis.call('HINCRBY', groupHashKey, 'tbTokens', -jobCostVal)
+    end
+    if rateMax > 0 then
+      local rateDuration = tonumber(grp.rateDuration) or 0
+      if rateDuration > 0 then
+        local rateWindowStart = tonumber(grp.rateWindowStart) or 0
+        local now = ts
+        if now - rateWindowStart >= rateDuration then
+          redis.call('HSET', groupHashKey, 'rateWindowStart', tostring(now), 'rateCount', '1')
+        else
+          redis.call('HINCRBY', groupHashKey, 'rateCount', 1)
+        end
+      end
+    end
+    if not isReturningStepJob then
+      redis.call('HINCRBY', groupHashKey, 'active', 1)
+    end
+    if jobOrderingSeq > 0 then
+      if not isReturningStepJob then
+        redis.call('HSET', groupHashKey, 'nextSeq', tostring(jobOrderingSeq + 1))
+      end
+      -- Remove from groupq in case job was parked earlier and re-delivered via priority list
+      redis.call('ZREM', waitListKey, jobId)
+    end
+  end
+  redis.call('HSET', jobKey, 'state', 'active', 'processedOn', timestampStr, 'lastActive', timestampStr)
+  if stateValueIndex ~= nil then
+    fields[stateValueIndex] = 'active'
+  else
+    fields[#fields + 1] = 'state'
+    fields[#fields + 1] = 'active'
+  end
+  if processedOnValueIndex ~= nil then
+    fields[processedOnValueIndex] = timestampStr
+  else
+    fields[#fields + 1] = 'processedOn'
+    fields[#fields + 1] = timestampStr
+  end
+  if lastActiveValueIndex ~= nil then
+    fields[lastActiveValueIndex] = timestampStr
+  else
+    fields[#fields + 1] = 'lastActive'
+    fields[#fields + 1] = timestampStr
+  end
+  return fields
 end)
 
--- glidemq_deferActive(KEYS, ARGS)
--- Defers an active job back to waiting by replacing its stream entry.
--- KEYS[1] = stream key
--- KEYS[2] = job hash key
--- ARGS[1] = job ID
--- ARGS[2] = stream entry ID
--- ARGS[3] = consumer group name
 redis.register_function('glidemq_deferActive', function(keys, args)
   local streamKey = keys[1]
   local jobKey = keys[2]
+  local listActiveKey = keys[3] or ''
   local jobId = args[1]
   local entryId = args[2]
   local group = args[3]
+  local broadcastMode = args[4] or '0'
   local exists = redis.call('EXISTS', jobKey)
-
-  redis.call('XACK', streamKey, group, entryId)
-  redis.call('XDEL', streamKey, entryId)
+  if entryId ~= '' then redis.call('XACK', streamKey, group, entryId) end
+  if entryId ~= '' and broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
+  -- List-sourced jobs (entryId='') were counted in list-active; DECR to stay balanced.
+  if entryId == '' then decrListActive(listActiveKey) end
   if exists == 0 then
     return 0
   end
-  redis.call('XADD', streamKey, '*', 'jobId', jobId)
+  xaddJob(streamKey, jobId, redis.call('HGET', jobKey, 'name'))
   redis.call('HSET', jobKey, 'state', 'waiting')
   return 1
 end)
 
--- Returns: cjson array [parentId, childId1, childId2, ...], or ['duplicate'] / ['ERR:ID_EXHAUSTED']
 redis.register_function('glidemq_addFlow', function(keys, args)
   local parentIdKey = keys[1]
   local parentStreamKey = keys[2]
   local parentScheduledKey = keys[3]
   local parentEventsKey = keys[4]
-
   local parentName = args[1]
   local parentData = args[2]
   local parentOpts = args[3]
@@ -951,11 +2304,7 @@ redis.register_function('glidemq_addFlow', function(keys, args)
   local parentMaxAttempts = tonumber(args[7]) or 0
   local numChildren = tonumber(args[8])
   local parentCustomId = args[9] or ''
-
-  -- Derive parent prefix from id key: "glide:{q}:id" -> "glide:{q}:"
   local parentPrefix = string.sub(parentIdKey, 1, #parentIdKey - 2)
-
-  -- Create parent job ID
   local parentJobIdStr
   local parentJobKey
   if parentCustomId ~= '' then
@@ -964,10 +2313,12 @@ redis.register_function('glidemq_addFlow', function(keys, args)
       return cjson.encode({'duplicate'})
     end
     parentJobIdStr = parentCustomId
+    advanceIdCounter(parentIdKey, parentCustomId)
   else
     local parentJobId = redis.call('INCR', parentIdKey)
     parentJobIdStr = tostring(parentJobId)
     parentJobKey = parentPrefix .. 'job:' .. parentJobIdStr
+    -- Skip auto-INCR'd IDs that collide with previously-created custom-ID jobs.
     local retries = 0
     while redis.call('EXISTS', parentJobKey) == 1 do
       retries = retries + 1
@@ -998,13 +2349,16 @@ redis.register_function('glidemq_addFlow', function(keys, args)
   end
   local depsKey = parentPrefix .. 'deps:' .. parentJobIdStr
   local parentOrderingKey = extractOrderingKeyFromOpts(parentOpts)
+  local parentGroupConc = extractGroupConcurrencyFromOpts(parentOpts)
+  local parentRateMax, parentRateDuration = extractGroupRateLimitFromOpts(parentOpts)
+  local parentTbCapacity, parentTbRefillRate = extractTokenBucketFromOpts(parentOpts)
+  local parentCost = extractCostFromOpts(parentOpts)
+  local parentUseGroup = (parentOrderingKey ~= '')
   local parentOrderingSeq = 0
   if parentOrderingKey ~= '' then
     local parentOrderingMetaKey = parentPrefix .. 'ordering'
     parentOrderingSeq = redis.call('HINCRBY', parentOrderingMetaKey, parentOrderingKey, 1)
   end
-
-  -- Store parent job hash with state=waiting-children
   local parentHash = {
     'id', parentJobIdStr,
     'name', parentName,
@@ -1017,20 +2371,60 @@ redis.register_function('glidemq_addFlow', function(keys, args)
     'maxAttempts', tostring(parentMaxAttempts),
     'state', 'waiting-children'
   }
-  if parentOrderingKey ~= '' then
-    parentHash[#parentHash + 1] = 'orderingKey'
+  if parentUseGroup then
+    parentHash[#parentHash + 1] = 'groupKey'
     parentHash[#parentHash + 1] = parentOrderingKey
-    parentHash[#parentHash + 1] = 'orderingSeq'
-    parentHash[#parentHash + 1] = tostring(parentOrderingSeq)
+    if parentOrderingSeq > 0 then
+      parentHash[#parentHash + 1] = 'orderingSeq'
+      parentHash[#parentHash + 1] = tostring(parentOrderingSeq)
+    end
+    if parentGroupConc < 1 then parentGroupConc = 1 end
+    local groupHashKey = parentPrefix .. 'group:' .. parentOrderingKey
+    redis.call('HSET', groupHashKey, 'maxConcurrency', tostring(parentGroupConc))
+    redis.call('HSETNX', groupHashKey, 'active', '0')
+    if parentOrderingSeq > 0 and redis.call('HEXISTS', groupHashKey, 'nextSeq') == 0 then
+      redis.call('HSET', groupHashKey, 'nextSeq', '1')
+    end
+    if parentRateMax > 0 then
+      redis.call('HSET', groupHashKey, 'rateMax', tostring(parentRateMax))
+      redis.call('HSET', groupHashKey, 'rateDuration', tostring(parentRateDuration))
+    end
+    if parentTbCapacity > 0 then
+      if parentCost > 0 and parentCost > parentTbCapacity then
+        return 'ERR:COST_EXCEEDS_CAPACITY'
+      end
+      redis.call('HSET', groupHashKey, 'tbCapacity', tostring(parentTbCapacity), 'tbRefillRate', tostring(parentTbRefillRate))
+      redis.call('HSETNX', groupHashKey, 'tbTokens', tostring(parentTbCapacity))
+      redis.call('HSETNX', groupHashKey, 'tbLastRefill', tostring(timestamp))
+      redis.call('HSETNX', groupHashKey, 'tbRefillRemainder', '0')
+    end
+  end
+  if parentCost > 0 then
+    parentHash[#parentHash + 1] = 'cost'
+    parentHash[#parentHash + 1] = tostring(parentCost)
+  end
+  local parentTtl = extractTtlFromOpts(parentOpts)
+  if parentTtl > 0 then
+    parentHash[#parentHash + 1] = 'expireAt'
+    parentHash[#parentHash + 1] = tostring(timestamp + parentTtl)
   end
   redis.call('HSET', parentJobKey, unpack(parentHash))
-
-  -- Collect child IDs for result
-  local childIds = {}
-
-  -- Process each child
+  -- Pre-validate all children's cost vs capacity before any child writes
   local childArgOffset = 9
   local childKeyOffset = 4
+  for i = 1, numChildren do
+    local base = childArgOffset + (i - 1) * 9
+    local preChildOpts = args[base + 3]
+    local preChildTbCap, _ = extractTokenBucketFromOpts(preChildOpts)
+    if preChildTbCap > 0 then
+      local preChildCost = extractCostFromOpts(preChildOpts)
+      local preEffective = (preChildCost > 0) and preChildCost or 1000
+      if preEffective > preChildTbCap then
+        return 'ERR:COST_EXCEEDS_CAPACITY'
+      end
+    end
+  end
+  local childIds = {}
   for i = 1, numChildren do
     local base = childArgOffset + (i - 1) * 9
     local childName = args[base + 1]
@@ -1042,18 +2436,12 @@ redis.register_function('glidemq_addFlow', function(keys, args)
     local childQueuePrefix = args[base + 7]
     local childParentQueue = args[base + 8]
     local childCustomId = args[base + 9] or ''
-
-    -- Child keys: use child-specific keys from KEYS array
     local ckBase = childKeyOffset + (i - 1) * 4
     local childIdKey = keys[ckBase + 1]
     local childStreamKey = keys[ckBase + 2]
     local childScheduledKey = keys[ckBase + 3]
     local childEventsKey = keys[ckBase + 4]
-
-    -- Derive child prefix from child id key
     local childPrefix = string.sub(childIdKey, 1, #childIdKey - 2)
-
-    -- Create child job ID (custom IDs pre-validated above)
     local childJobIdStr
     local childJobKey
     if childCustomId ~= '' then
@@ -1062,27 +2450,34 @@ redis.register_function('glidemq_addFlow', function(keys, args)
         return cjson.encode({'duplicate'})
       end
       childJobIdStr = childCustomId
+      advanceIdCounter(childIdKey, childCustomId)
     else
       local childJobId = redis.call('INCR', childIdKey)
       childJobIdStr = tostring(childJobId)
       childJobKey = childPrefix .. 'job:' .. childJobIdStr
-      local cRetries = 0
+      -- Skip over IDs that collide with custom-ID jobs (parent or sibling),
+      -- mirroring glidemq_addJob's auto-ID collision guard.
+      local retries = 0
       while redis.call('EXISTS', childJobKey) == 1 do
-        cRetries = cRetries + 1
-        if cRetries >= 1000 then return cjson.encode({'ERR:ID_EXHAUSTED'}) end
+        retries = retries + 1
+        if retries >= 1000 then return cjson.encode({'ERR:ID_EXHAUSTED'}) end
         childJobId = redis.call('INCR', childIdKey)
         childJobIdStr = tostring(childJobId)
         childJobKey = childPrefix .. 'job:' .. childJobIdStr
       end
     end
     local childOrderingKey = extractOrderingKeyFromOpts(childOpts)
+    local childLifo = extractLifoFromOpts(childOpts)
+    local childGroupConc = extractGroupConcurrencyFromOpts(childOpts)
+    local childRateMax, childRateDuration = extractGroupRateLimitFromOpts(childOpts)
+    local childTbCapacity, childTbRefillRate = extractTokenBucketFromOpts(childOpts)
+    local childCost = extractCostFromOpts(childOpts)
+    local childUseGroup = (childOrderingKey ~= '')
     local childOrderingSeq = 0
     if childOrderingKey ~= '' then
       local childOrderingMetaKey = childPrefix .. 'ordering'
       childOrderingSeq = redis.call('HINCRBY', childOrderingMetaKey, childOrderingKey, 1)
     end
-
-    -- Store child job hash with parentId and parentQueue
     local childHash = {
       'id', childJobIdStr,
       'name', childName,
@@ -1096,13 +2491,44 @@ redis.register_function('glidemq_addFlow', function(keys, args)
       'parentId', parentJobIdStr,
       'parentQueue', childParentQueue
     }
-    if childOrderingKey ~= '' then
-      childHash[#childHash + 1] = 'orderingKey'
+    if childUseGroup then
+      childHash[#childHash + 1] = 'groupKey'
       childHash[#childHash + 1] = childOrderingKey
-      childHash[#childHash + 1] = 'orderingSeq'
-      childHash[#childHash + 1] = tostring(childOrderingSeq)
+      if childOrderingSeq > 0 then
+        childHash[#childHash + 1] = 'orderingSeq'
+        childHash[#childHash + 1] = tostring(childOrderingSeq)
+      end
+      if childGroupConc < 1 then childGroupConc = 1 end
+      local childGroupHashKey = childPrefix .. 'group:' .. childOrderingKey
+      redis.call('HSETNX', childGroupHashKey, 'maxConcurrency', tostring(childGroupConc))
+      redis.call('HSETNX', childGroupHashKey, 'active', '0')
+      if childOrderingSeq > 0 and redis.call('HEXISTS', childGroupHashKey, 'nextSeq') == 0 then
+        redis.call('HSET', childGroupHashKey, 'nextSeq', '1')
+      end
+      if childRateMax > 0 then
+        redis.call('HSET', childGroupHashKey, 'rateMax', tostring(childRateMax))
+        redis.call('HSET', childGroupHashKey, 'rateDuration', tostring(childRateDuration))
+      end
+      if childTbCapacity > 0 then
+        redis.call('HSET', childGroupHashKey, 'tbCapacity', tostring(childTbCapacity), 'tbRefillRate', tostring(childTbRefillRate))
+        redis.call('HSETNX', childGroupHashKey, 'tbTokens', tostring(childTbCapacity))
+        redis.call('HSETNX', childGroupHashKey, 'tbLastRefill', tostring(timestamp))
+        redis.call('HSETNX', childGroupHashKey, 'tbRefillRemainder', '0')
+      end
     end
-
+    if childCost > 0 then
+      childHash[#childHash + 1] = 'cost'
+      childHash[#childHash + 1] = tostring(childCost)
+    end
+    local childTtl = extractTtlFromOpts(childOpts)
+    if childTtl > 0 then
+      childHash[#childHash + 1] = 'expireAt'
+      childHash[#childHash + 1] = tostring(timestamp + childTtl)
+    end
+    if childLifo > 0 then
+      childHash[#childHash + 1] = 'lifo'
+      childHash[#childHash + 1] = '1'
+    end
     if childDelay > 0 or childPriority > 0 then
       childHash[#childHash + 1] = 'state'
       childHash[#childHash + 1] = childDelay > 0 and 'delayed' or 'prioritized'
@@ -1110,39 +2536,30 @@ redis.register_function('glidemq_addFlow', function(keys, args)
       childHash[#childHash + 1] = 'state'
       childHash[#childHash + 1] = 'waiting'
     end
-
     redis.call('HSET', childJobKey, unpack(childHash))
-
-    -- Add child to deps set (using queuePrefix:childId as identifier for cross-queue tracking)
     local depsMember = childQueuePrefix .. ':' .. childJobIdStr
     redis.call('SADD', depsKey, depsMember)
-
-    -- Route child job
     if childDelay > 0 then
       local score = childPriority * PRIORITY_SHIFT + (timestamp + childDelay)
       redis.call('ZADD', childScheduledKey, score, childJobIdStr)
     elseif childPriority > 0 then
       local score = childPriority * PRIORITY_SHIFT
       redis.call('ZADD', childScheduledKey, score, childJobIdStr)
+    elseif childLifo > 0 then
+      redis.call('RPUSH', childPrefix .. 'lifo', childJobIdStr)
     else
-      redis.call('XADD', childStreamKey, '*', 'jobId', childJobIdStr)
+      xaddJob(childStreamKey, childJobIdStr, childName)
     end
-
     emitEvent(childEventsKey, 'added', childJobIdStr, {'name', childName})
     childIds[#childIds + 1] = childJobIdStr
   end
-
-  -- Process extra deps members (pre-existing sub-flow children)
   local extraDepsOffset = childArgOffset + numChildren * 9
   local numExtraDeps = tonumber(args[extraDepsOffset + 1]) or 0
   for i = 1, numExtraDeps do
     local extraMember = args[extraDepsOffset + 1 + i]
     redis.call('SADD', depsKey, extraMember)
   end
-
   emitEvent(parentEventsKey, 'added', parentJobIdStr, {'name', parentName})
-
-  -- Return JSON array: [parentId, child1Id, child2Id, ...]
   local result = {parentJobIdStr}
   for i = 1, #childIds do
     result[#result + 1] = childIds[i]
@@ -1150,51 +2567,76 @@ redis.register_function('glidemq_addFlow', function(keys, args)
   return cjson.encode(result)
 end)
 
--- glidemq_completeChild(KEYS, ARGS)
--- Removes a child from the parent's deps set. If all children are done, re-queues the parent.
--- KEYS[1] = parent deps key (glide:{parentQueue}:deps:{parentId})
--- KEYS[2] = parent job hash key (glide:{parentQueue}:job:{parentId})
--- KEYS[3] = parent stream key (glide:{parentQueue}:stream)
--- KEYS[4] = parent events key (glide:{parentQueue}:events)
--- ARGS[1] = deps member string (childQueuePrefix:childId)
--- ARGS[2] = parent job ID
--- Returns: number of remaining children (0 = parent re-queued)
 redis.register_function('glidemq_completeChild', function(keys, args)
   local depsKey = keys[1]
   local parentJobKey = keys[2]
   local parentStreamKey = keys[3]
   local parentEventsKey = keys[4]
-
   local depsMember = args[1]
   local parentId = args[2]
-
-  -- Increment completed children counter on parent hash
+  local depMarker = 'depdone:' .. depsMember
+  if redis.call('HSETNX', parentJobKey, depMarker, '1') == 0 then
+    local doneCount = tonumber(redis.call('HGET', parentJobKey, 'depsCompleted')) or 0
+    local totalDeps = redis.call('SCARD', depsKey)
+    return totalDeps - doneCount
+  end
   local doneCount = redis.call('HINCRBY', parentJobKey, 'depsCompleted', 1)
-
-  -- Compare with total deps count
   local totalDeps = redis.call('SCARD', depsKey)
   local remaining = totalDeps - doneCount
-
   if remaining <= 0 then
-    -- All children done: move parent to stream (re-queue)
-    redis.call('HSET', parentJobKey, 'state', 'waiting')
-    redis.call('XADD', parentStreamKey, '*', 'jobId', parentId)
-    emitEvent(parentEventsKey, 'active', parentId, nil)
+    local parentState = redis.call('HGET', parentJobKey, 'state')
+    if parentState == 'waiting-children' then
+      redis.call('HSET', parentJobKey, 'state', 'waiting')
+      xaddJob(parentStreamKey, parentId, redis.call('HGET', parentJobKey, 'name'))
+      emitEvent(parentEventsKey, 'active', parentId, nil)
+    end
   end
-
   return remaining
 end)
 
--- glidemq_removeJob(KEYS, ARGS)
--- Removes a job from all data structures.
--- KEYS[1] = job hash key (job:{id})
--- KEYS[2] = stream key
--- KEYS[3] = scheduled zset key
--- KEYS[4] = completed zset key
--- KEYS[5] = failed zset key
--- KEYS[6] = events stream key
--- ARGS[1] = job ID
--- Returns: 1 if removed, 0 if not found
+redis.register_function('glidemq_registerParent', function(keys, args)
+  -- Register an additional parent for an existing child job (DAG multi-parent).
+  -- Keys: [childJobKey, childParentsKey, parentDepsKey, parentJobKey, parentStreamKey, parentEventsKey]
+  -- Args: [childJobId, parentId, parentQueue, depsMember]
+  local childJobKey = keys[1]
+  local childParentsKey = keys[2]
+  local parentDepsKey = keys[3]
+  local parentJobKey = keys[4]
+  local parentStreamKey = keys[5]
+  local parentEventsKey = keys[6]
+  local childJobId = args[1]
+  local parentId = args[2]
+  local parentQueue = args[3]
+  local depsMember = args[4]
+  -- Verify child exists
+  if redis.call('EXISTS', childJobKey) == 0 then
+    return 'error:child_not_found'
+  end
+  -- Add parent entry to child's parents SET (idempotent)
+  redis.call('SADD', childParentsKey, parentQueue .. ':' .. parentId)
+  -- Add child as dependency in parent's deps SET (idempotent)
+  redis.call('SADD', parentDepsKey, depsMember)
+  -- Race condition check: if child already completed, trigger parent notification immediately
+  local childState = redis.call('HGET', childJobKey, 'state')
+  if childState == 'completed' then
+    local depMarker = 'depdone:' .. depsMember
+    if redis.call('HSETNX', parentJobKey, depMarker, '1') == 1 then
+      local doneCount = redis.call('HINCRBY', parentJobKey, 'depsCompleted', 1)
+      local totalDeps = redis.call('SCARD', parentDepsKey)
+      if totalDeps - doneCount <= 0 then
+        local parentState = redis.call('HGET', parentJobKey, 'state')
+        if parentState == 'waiting-children' then
+          redis.call('HSET', parentJobKey, 'state', 'waiting')
+          xaddJob(parentStreamKey, parentId, redis.call('HGET', parentJobKey, 'name'))
+          emitEvent(parentEventsKey, 'active', parentId, nil)
+        end
+      end
+    end
+    return 'already_completed'
+  end
+  return 'ok'
+end)
+
 redis.register_function('glidemq_removeJob', function(keys, args)
   local jobKey = keys[1]
   local streamKey = keys[2]
@@ -1202,26 +2644,1075 @@ redis.register_function('glidemq_removeJob', function(keys, args)
   local completedKey = keys[4]
   local failedKey = keys[5]
   local eventsKey = keys[6]
-
+  local logKey = keys[7]
   local jobId = args[1]
-
-  -- Check if job exists
   local exists = redis.call('EXISTS', jobKey)
   if exists == 0 then
     return 0
   end
-
-  -- Remove from scheduled, completed, failed ZSets
+  local state = redis.call('HGET', jobKey, 'state')
+  local groupKey = redis.call('HGET', jobKey, 'groupKey')
+  if groupKey and groupKey ~= '' then
+    if state == 'active' then
+      releaseGroupSlotAndPromote(jobKey, jobId, 0)
+    elseif state == 'group-waiting' then
+      local prefix = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
+      local waitListKey = prefix .. 'groupq:' .. groupKey
+      redis.call('ZREM', waitListKey, jobId)
+    end
+  end
+  -- DECR list-active if the job was active and list-sourced (LIFO or priority-list)
+  if state == 'active' then
+    local jobLifo = redis.call('HGET', jobKey, 'lifo')
+    local jobPriority = tonumber(redis.call('HGET', jobKey, 'priority')) or 0
+    if jobLifo == '1' or jobPriority > 0 then
+      local prefix_r = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
+      decrListActive(prefix_r .. 'list-active')
+    end
+  end
   redis.call('ZREM', scheduledKey, jobId)
   redis.call('ZREM', completedKey, jobId)
   redis.call('ZREM', failedKey, jobId)
   markOrderingDone(jobKey, jobId)
-
-  -- Delete the job hash
-  redis.call('UNLINK', jobKey)
-
-  -- Emit event
+  -- Clean up DAG parents SET, per-job streaming channel, and signals.
+  -- Job hash + log can be MB-sized; parents/jstream/signals carry per-step
+  -- data. Use UNLINK so the server reclaims memory off the main thread.
+  local prefix = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
+  local parentsKey = prefix .. 'parents:' .. jobId
+  local jstreamKey = prefix .. 'jstream:' .. jobId
+  local signalsKey = prefix .. 'signals:' .. jobId
+  local suspendedKey = prefix .. 'suspended'
+  redis.call('UNLINK', parentsKey, jstreamKey, signalsKey, jobKey, logKey)
+  redis.call('ZREM', suspendedKey, jobId)
   emitEvent(eventsKey, 'removed', jobId, nil)
-
   return 1
+end)
+
+redis.register_function('glidemq_clean', function(keys, args)
+  local setKey = keys[1]
+  local eventsKey = keys[2]
+  local idKey = keys[3]
+  local cutoff = tonumber(args[1])
+  local limit = tonumber(args[2])
+  if not limit or limit <= 0 then return {} end
+  local prefix = string.sub(idKey, 1, #idKey - 2)
+  local ids = redis.call('ZRANGEBYSCORE', setKey, '-inf', string.format('%.0f', cutoff), 'LIMIT', 0, limit)
+  if #ids == 0 then
+    return {}
+  end
+  for i = 1, #ids do
+    redis.call('UNLINK', prefix .. 'job:' .. ids[i], prefix .. 'log:' .. ids[i], prefix .. 'deps:' .. ids[i], prefix .. 'parents:' .. ids[i], prefix .. 'jstream:' .. ids[i], prefix .. 'signals:' .. ids[i])
+  end
+  for i = 1, #ids, 1000 do
+    redis.call('ZREM', setKey, unpack(ids, i, math.min(i + 999, #ids)))
+  end
+  emitEvent(eventsKey, 'cleaned', tostring(#ids), nil)
+  return ids
+end)
+
+redis.register_function('glidemq_revoke', function(keys, args)
+  local jobKey = keys[1]
+  local streamKey = keys[2]
+  local scheduledKey = keys[3]
+  local failedKey = keys[4]
+  local eventsKey = keys[5]
+  local jobId = args[1]
+  local timestamp = tonumber(args[2])
+  local group = args[3]
+  local exists = redis.call('EXISTS', jobKey)
+  if exists == 0 then
+    return 'not_found'
+  end
+  redis.call('HSET', jobKey, 'revoked', '1')
+  local state = redis.call('HGET', jobKey, 'state')
+  if state == 'group-waiting' then
+    local gk = redis.call('HGET', jobKey, 'groupKey')
+    if gk and gk ~= '' then
+      local prefix = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
+      local waitListKey = prefix .. 'groupq:' .. gk
+      redis.call('ZREM', waitListKey, jobId)
+    end
+    redis.call('ZADD', failedKey, timestamp, jobId)
+    redis.call('HSET', jobKey,
+      'state', 'failed',
+      'failedReason', 'revoked',
+      'finishedOn', tostring(timestamp)
+    )
+    emitEvent(eventsKey, 'revoked', jobId, nil)
+    return 'revoked'
+  end
+  if state == 'waiting' or state == 'delayed' or state == 'prioritized' then
+    redis.call('ZREM', scheduledKey, jobId)
+    local cursor = '-'
+    local found = false
+    while not found do
+      local entries = redis.call('XRANGE', streamKey, cursor, '+', 'COUNT', 1000)
+      if #entries == 0 then break end
+      for i = 1, #entries do
+        local entryId = entries[i][1]
+        local fields = entries[i][2]
+        for j = 1, #fields, 2 do
+          if fields[j] == 'jobId' and fields[j+1] == jobId then
+            if entryId ~= '' then redis.call('XACK', streamKey, group, entryId) end
+            if entryId ~= '' then redis.call('XDEL', streamKey, entryId) end
+            found = true
+            break
+          end
+        end
+        if found then break end
+      end
+      if not found then
+        local lastId = entries[#entries][1]
+        local dashPos = lastId:find('-')
+        cursor = lastId:sub(1, dashPos) .. tostring(tonumber(lastId:sub(dashPos + 1)) + 1)
+      end
+    end
+    redis.call('ZADD', failedKey, timestamp, jobId)
+    redis.call('HSET', jobKey,
+      'state', 'failed',
+      'failedReason', 'revoked',
+      'finishedOn', tostring(timestamp)
+    )
+    markOrderingDone(jobKey, jobId)
+    emitEvent(eventsKey, 'revoked', jobId, nil)
+    return 'revoked'
+  end
+  emitEvent(eventsKey, 'revoked', jobId, nil)
+  return 'flagged'
+end)
+
+redis.register_function('glidemq_changePriority', function(keys, args)
+  local jobKey = keys[1]
+  local streamKey = keys[2]
+  local scheduledKey = keys[3]
+  local eventsKey = keys[4]
+  local jobId = args[1]
+  local newPriority = tonumber(args[2])
+  if newPriority == nil or newPriority < 0 then
+    return 'error:invalid_priority'
+  end
+  local group = args[3]
+  local exists = redis.call('EXISTS', jobKey)
+  if exists == 0 then
+    return 'error:not_found'
+  end
+  local state = redis.call('HGET', jobKey, 'state')
+  if state == 'waiting' then
+    if newPriority == 0 then
+      return 'no_op'
+    end
+    local cursor = '-'
+    local found = false
+    while not found do
+      local entries = redis.call('XRANGE', streamKey, cursor, '+', 'COUNT', 1000)
+      if #entries == 0 then break end
+      for i = 1, #entries do
+        local entryId = entries[i][1]
+        local fields = entries[i][2]
+        for j = 1, #fields, 2 do
+          if fields[j] == 'jobId' and fields[j+1] == jobId then
+            pcall(redis.call, 'XACK', streamKey, group, entryId)
+            if entryId ~= '' then redis.call('XDEL', streamKey, entryId) end
+            found = true
+            break
+          end
+        end
+        if found then break end
+      end
+      if not found then
+        local lastId = entries[#entries][1]
+        local dashPos = lastId:find('-')
+        cursor = lastId:sub(1, dashPos) .. tostring(tonumber(lastId:sub(dashPos + 1)) + 1)
+      end
+    end
+    if not found then
+      return 'error:not_in_stream'
+    end
+    redis.call('ZADD', scheduledKey, string.format('%.0f', newPriority * PRIORITY_SHIFT), jobId)
+    redis.call('HSET', jobKey, 'state', 'prioritized', 'priority', tostring(newPriority))
+    emitEvent(eventsKey, 'priority-changed', jobId, {'priority', tostring(newPriority)})
+    return 'ok'
+  elseif state == 'prioritized' then
+    if newPriority == 0 then
+      redis.call('ZREM', scheduledKey, jobId)
+      xaddJob(streamKey, jobId, redis.call('HGET', jobKey, 'name'))
+      redis.call('HSET', jobKey, 'state', 'waiting', 'priority', '0')
+    else
+      redis.call('ZADD', scheduledKey, string.format('%.0f', newPriority * PRIORITY_SHIFT), jobId)
+      redis.call('HSET', jobKey, 'priority', tostring(newPriority))
+    end
+    emitEvent(eventsKey, 'priority-changed', jobId, {'priority', tostring(newPriority)})
+    return 'ok'
+  elseif state == 'delayed' then
+    local rawScore = redis.call('ZSCORE', scheduledKey, jobId)
+    if rawScore == false then
+      return 'error:not_in_scheduled'
+    end
+    local oldScore = tonumber(rawScore) or 0
+    local oldTimestamp = oldScore % PRIORITY_SHIFT
+    local newScore = newPriority * PRIORITY_SHIFT + oldTimestamp
+    redis.call('ZREM', scheduledKey, jobId)
+    redis.call('ZADD', scheduledKey, string.format('%.0f', newScore), jobId)
+    redis.call('HSET', jobKey, 'priority', tostring(newPriority))
+    emitEvent(eventsKey, 'priority-changed', jobId, {'priority', tostring(newPriority)})
+    return 'ok'
+  else
+    return 'error:invalid_state'
+  end
+end)
+
+redis.register_function('glidemq_changeDelay', function(keys, args)
+  local jobKey = keys[1]
+  local streamKey = keys[2]
+  local scheduledKey = keys[3]
+  local eventsKey = keys[4]
+  local jobId = args[1]
+  local newDelay = tonumber(args[2])
+  if newDelay == nil or newDelay < 0 then
+    return 'error:invalid_delay'
+  end
+  local now = tonumber(args[3])
+  local group = args[4]
+  local exists = redis.call('EXISTS', jobKey)
+  if exists == 0 then
+    return 'error:not_found'
+  end
+  local state = redis.call('HGET', jobKey, 'state')
+  if state == 'delayed' then
+    if newDelay == 0 then
+      local rawScore = redis.call('ZSCORE', scheduledKey, jobId)
+      if rawScore == false then
+        return 'error:not_in_scheduled'
+      end
+      local oldScore = tonumber(rawScore) or 0
+      local priority = math.floor(oldScore / PRIORITY_SHIFT)
+      if priority > 0 then
+        redis.call('ZADD', scheduledKey, 'XX', string.format('%.0f', priority * PRIORITY_SHIFT), jobId)
+        redis.call('HSET', jobKey, 'state', 'prioritized', 'delay', '0')
+      else
+        redis.call('ZREM', scheduledKey, jobId)
+        xaddJob(streamKey, jobId, redis.call('HGET', jobKey, 'name'))
+        redis.call('HSET', jobKey, 'state', 'waiting', 'delay', '0')
+      end
+    else
+      local rawScore = redis.call('ZSCORE', scheduledKey, jobId)
+      if rawScore == false then
+        return 'error:not_in_scheduled'
+      end
+      local oldScore = tonumber(rawScore) or 0
+      local priority = math.floor(oldScore / PRIORITY_SHIFT)
+      local newScore = priority * PRIORITY_SHIFT + (now + newDelay)
+      redis.call('ZADD', scheduledKey, 'XX', string.format('%.0f', newScore), jobId)
+      redis.call('HSET', jobKey, 'delay', tostring(newDelay))
+    end
+    emitEvent(eventsKey, 'delay-changed', jobId, {'delay', tostring(newDelay)})
+    return 'ok'
+  elseif state == 'waiting' then
+    if newDelay == 0 then
+      return 'no_op'
+    end
+    local priority = tonumber(redis.call('HGET', jobKey, 'priority')) or 0
+    local cursor = '-'
+    local found = false
+    while not found do
+      local entries = redis.call('XRANGE', streamKey, cursor, '+', 'COUNT', 1000)
+      if #entries == 0 then break end
+      for i = 1, #entries do
+        local entryId = entries[i][1]
+        local fields = entries[i][2]
+        for j = 1, #fields, 2 do
+          if fields[j] == 'jobId' and fields[j+1] == jobId then
+            pcall(redis.call, 'XACK', streamKey, group, entryId)
+            if entryId ~= '' then redis.call('XDEL', streamKey, entryId) end
+            found = true
+            break
+          end
+        end
+        if found then break end
+      end
+      if not found then
+        cursor = '(' .. entries[#entries][1]
+      end
+    end
+    if not found then
+      return 'error:not_in_stream'
+    end
+    local newScore = priority * PRIORITY_SHIFT + (now + newDelay)
+    redis.call('ZADD', scheduledKey, string.format('%.0f', newScore), jobId)
+    redis.call('HSET', jobKey, 'state', 'delayed', 'delay', tostring(newDelay))
+    emitEvent(eventsKey, 'delay-changed', jobId, {'delay', tostring(newDelay)})
+    return 'ok'
+  elseif state == 'prioritized' then
+    if newDelay == 0 then
+      return 'no_op'
+    end
+    local rawScore = redis.call('ZSCORE', scheduledKey, jobId)
+    if rawScore == false then
+      return 'error:not_in_scheduled'
+    end
+    local oldScore = tonumber(rawScore) or 0
+    local priority = math.floor(oldScore / PRIORITY_SHIFT)
+    local newScore = priority * PRIORITY_SHIFT + (now + newDelay)
+    redis.call('ZADD', scheduledKey, 'XX', string.format('%.0f', newScore), jobId)
+    redis.call('HSET', jobKey, 'state', 'delayed', 'delay', tostring(newDelay))
+    emitEvent(eventsKey, 'delay-changed', jobId, {'delay', tostring(newDelay)})
+    return 'ok'
+  else
+    return 'error:invalid_state'
+  end
+end)
+
+redis.register_function('glidemq_promoteJob', function(keys, args)
+  local jobKey = keys[1]
+  local streamKey = keys[2]
+  local scheduledKey = keys[3]
+  local eventsKey = keys[4]
+  local jobId = args[1]
+  local exists = redis.call('EXISTS', jobKey)
+  if exists == 0 then
+    return 'error:not_found'
+  end
+  local state = redis.call('HGET', jobKey, 'state')
+  if state ~= 'delayed' then
+    return 'error:not_delayed'
+  end
+  redis.call('ZREM', scheduledKey, jobId)
+  local jobPriority = tonumber(redis.call('HGET', jobKey, 'priority')) or 0
+  local jobLifo = redis.call('HGET', jobKey, 'lifo')
+  local prefix = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
+  if jobLifo == '1' then
+    redis.call('RPUSH', prefix .. 'lifo', jobId)
+  elseif jobPriority > 0 then
+    redis.call('LPUSH', prefix .. 'priority', jobId)
+  else
+    xaddJob(streamKey, jobId, redis.call('HGET', jobKey, 'name'))
+  end
+  redis.call('HSET', jobKey, 'state', 'waiting', 'delay', '0')
+  emitEvent(eventsKey, 'promoted', jobId, nil)
+  return 'ok'
+end)
+
+redis.register_function('glidemq_moveActiveToDelayed', function(keys, args)
+  local jobKey = keys[1]
+  local streamKey = keys[2]
+  local scheduledKey = keys[3]
+  local eventsKey = keys[4]
+  local jobId = args[1]
+  local entryId = args[2]
+  local now = tonumber(args[3]) or 0
+  local delayedUntil = tonumber(args[4]) or now
+  local group = args[5]
+  local nextData = args[6]
+  local broadcastMode = args[7] or '0'
+
+  if redis.call('EXISTS', jobKey) == 0 then
+    return 'error:not_found'
+  end
+
+  local state = redis.call('HGET', jobKey, 'state')
+  if state ~= 'active' then
+    return 'error:not_active'
+  end
+
+  if delayedUntil < now then
+    delayedUntil = now
+  end
+
+  local priority = tonumber(redis.call('HGET', jobKey, 'priority')) or 0
+  local delay = delayedUntil - now
+  local score = priority * PRIORITY_SHIFT + delayedUntil
+
+  pcall(redis.call, 'XACK', streamKey, group, entryId)
+  if entryId ~= '' and broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
+  redis.call('ZADD', scheduledKey, string.format('%.0f', score), jobId)
+  if nextData and nextData ~= '' then
+    redis.call('HSET', jobKey, 'data', nextData, 'state', 'delayed', 'delay', tostring(delay))
+  else
+    redis.call('HSET', jobKey, 'state', 'delayed', 'delay', tostring(delay))
+  end
+  if entryId == '' then decrListActive(string.sub(jobKey, 1, #jobKey - #('job:' .. jobId)) .. 'list-active') end
+  -- Only release group slot if this is NOT an ordering-key step-job.
+  -- Ordering-key jobs hold the slot until full completion to preserve per-key order.
+  local jobOrdSeq = tonumber(redis.call('HGET', jobKey, 'orderingSeq')) or 0
+  if jobOrdSeq <= 0 then
+    releaseGroupSlotAndPromote(jobKey, jobId, now, nil)
+  end
+  emitEvent(eventsKey, 'delay-changed', jobId, {'delay', tostring(delay)})
+  return 'ok'
+end)
+
+redis.register_function('glidemq_moveToWaitingChildren', function(keys, args)
+  local jobKey = keys[1]
+  local streamKey = keys[2]
+  local eventsKey = keys[3]
+  local jobId = args[1]
+  local entryId = args[2]
+  local group = args[3]
+  local now = tonumber(args[4]) or 0
+  local broadcastMode = args[5] or '0'
+
+  local state = redis.call('HGET', jobKey, 'state')
+  if not state then
+    return 'error:not_found'
+  end
+  if state ~= 'active' then
+    return 'error:not_active'
+  end
+
+  pcall(redis.call, 'XACK', streamKey, group, entryId)
+  if entryId ~= '' and broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
+  redis.call('HSET', jobKey, 'state', 'waiting-children')
+
+  local wcOrdSeq = tonumber(redis.call('HGET', jobKey, 'orderingSeq')) or 0
+  if wcOrdSeq <= 0 then
+    releaseGroupSlotAndPromote(jobKey, jobId, now)
+  end
+
+  -- Race condition check: children may have already completed before this call
+  local prefix = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
+  local depsKey = prefix .. 'deps:' .. jobId
+  local totalDeps = redis.call('SCARD', depsKey)
+  if totalDeps > 0 then
+    local depsCompleted = tonumber(redis.call('HGET', jobKey, 'depsCompleted')) or 0
+    if depsCompleted >= totalDeps then
+      redis.call('HSET', jobKey, 'state', 'waiting')
+      xaddJob(streamKey, jobId, redis.call('HGET', jobKey, 'name'))
+      emitEvent(eventsKey, 'active', jobId, nil)
+      if entryId == '' then decrListActive(prefix .. 'list-active') end
+      return 'completed'
+    end
+  end
+
+  if entryId == '' then decrListActive(prefix .. 'list-active') end
+  emitEvent(eventsKey, 'waiting-children', jobId, nil)
+  return 'ok'
+end)
+
+redis.register_function('glidemq_rateLimitGroup', function(keys, args)
+  local jobKey = keys[1]
+  local streamKey = keys[2]
+  local jobId = args[1]
+  local entryId = args[2]
+  local group = args[3]
+  local duration = tonumber(args[4]) or 0
+  local timestamp = tonumber(args[5]) or 0
+  local broadcastMode = args[6] or '0'
+  local currentJob = args[7] or 'requeue'
+  local requeuePosition = args[8] or 'front'
+  local extend = args[9] or 'max'
+
+  local state = redis.call('HGET', jobKey, 'state')
+  if state ~= 'active' then return 'error:not_active' end
+
+  local groupKey = redis.call('HGET', jobKey, 'groupKey')
+  if not groupKey or groupKey == '' then return 'error:no_group' end
+
+  local prefix = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
+  local groupHashKey = prefix .. 'group:' .. groupKey
+  local waitListKey = prefix .. 'groupq:' .. groupKey
+  local rateLimitedKey = prefix .. 'ratelimited'
+  local eventsKey = prefix .. 'events'
+
+  if entryId ~= '' then pcall(redis.call, 'XACK', streamKey, group, entryId) end
+  if entryId ~= '' and broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
+  if entryId == '' then decrListActive(prefix .. 'list-active') end
+
+  local active = tonumber(redis.call('HGET', groupHashKey, 'active')) or 0
+  if active > 0 then redis.call('HSET', groupHashKey, 'active', tostring(active - 1)) end
+
+  if currentJob == 'requeue' then
+    local orderingSeq = tonumber(redis.call('HGET', jobKey, 'orderingSeq')) or 0
+    local score
+    if requeuePosition == 'front' then
+      score = orderingSeq > 0 and orderingSeq or 0
+    else
+      local lastEntry = redis.call('ZRANGE', waitListKey, -1, -1, 'WITHSCORES')
+      if lastEntry and #lastEntry >= 2 then
+        score = tonumber(lastEntry[2]) + 1
+      else
+        score = orderingSeq > 0 and orderingSeq or (tonumber(jobId) or 0)
+      end
+    end
+    redis.call('ZADD', waitListKey, score, jobId)
+    redis.call('HSET', jobKey, 'state', 'group-waiting')
+  else
+    redis.call('ZADD', prefix .. 'failed', timestamp, jobId)
+    local processedOn = tonumber(redis.call('HGET', jobKey, 'processedOn')) or timestamp
+    redis.call('HSET', jobKey, 'state', 'failed', 'failedReason', 'group rate limited', 'finishedOn', tostring(timestamp), 'processedOn', tostring(processedOn))
+    emitEvent(eventsKey, 'failed', jobId, {'failedReason', 'group rate limited'})
+    local metricsKey = prefix .. 'metrics:failed'
+    recordMetrics(metricsKey, timestamp, timestamp - processedOn)
+  end
+
+  local resumeAt = timestamp + duration
+  if extend == 'max' then
+    local existing = tonumber(redis.call('ZSCORE', rateLimitedKey, groupKey))
+    if existing and existing > resumeAt then resumeAt = existing end
+  end
+  redis.call('ZADD', rateLimitedKey, resumeAt, groupKey)
+
+  emitEvent(eventsKey, 'group-rate-limited', jobId, {
+    'groupKey', groupKey, 'duration', tostring(duration), 'resumeAt', tostring(resumeAt)
+  })
+  return tostring(resumeAt)
+end)
+
+redis.register_function('glidemq_rateLimitGroupExternal', function(keys, args)
+  local rateLimitedKey = keys[1]
+  local eventsKey = keys[2]
+  local groupKey = args[1]
+  local duration = tonumber(args[2]) or 0
+  local timestamp = tonumber(args[3]) or 0
+  local extend = args[4] or 'max'
+
+  local resumeAt = timestamp + duration
+  if extend == 'max' then
+    local existing = tonumber(redis.call('ZSCORE', rateLimitedKey, groupKey))
+    if existing and existing > resumeAt then resumeAt = existing end
+  end
+  redis.call('ZADD', rateLimitedKey, resumeAt, groupKey)
+
+  emitEvent(eventsKey, 'group-rate-limited', '', {
+    'groupKey', groupKey, 'duration', tostring(duration), 'resumeAt', tostring(resumeAt)
+  })
+  return resumeAt
+end)
+
+redis.register_function('glidemq_searchByName', function(keys, args)
+  local stateKey = keys[1]
+  local stateType = args[1]
+  local nameFilter = args[2]
+  local limit = tonumber(args[3]) or 100
+  local prefix = args[4]
+  local matched = {}
+  if stateType == 'zset' then
+    local members = redis.call('ZRANGE', stateKey, 0, -1)
+    for i = 1, #members do
+      if #matched >= limit then break end
+      local jobId = members[i]
+      local jobKey = prefix .. 'job:' .. jobId
+      local name = redis.call('HGET', jobKey, 'name')
+      if name == nameFilter then
+        matched[#matched + 1] = jobId
+      end
+    end
+  elseif stateType == 'stream' then
+    local cursor = '-'
+    while #matched < limit do
+      local entries = redis.call('XRANGE', stateKey, cursor, '+', 'COUNT', 1000)
+      if #entries == 0 then break end
+      for i = 1, #entries do
+        if #matched >= limit then break end
+        local fields = entries[i][2]
+        local jobId = nil
+        for j = 1, #fields, 2 do
+          if fields[j] == 'jobId' then
+            jobId = fields[j + 1]
+            break
+          end
+        end
+        if jobId then
+          local jobKey = prefix .. 'job:' .. jobId
+          local name = redis.call('HGET', jobKey, 'name')
+          if name == nameFilter then
+            matched[#matched + 1] = jobId
+          end
+        end
+      end
+      local lastId = entries[#entries][1]
+      local dashPos = lastId:find('-')
+      cursor = lastId:sub(1, dashPos) .. tostring(tonumber(lastId:sub(dashPos + 1)) + 1)
+    end
+  end
+  return matched
+end)
+
+redis.register_function('glidemq_drain', function(keys, args)
+  local streamKey = keys[1]
+  local scheduledKey = keys[2]
+  local eventsKey = keys[3]
+  local idKey = keys[4]
+  local lifoKey = keys[5]
+  local priorityKey = keys[6]
+  local drainDelayed = args[1] == '1'
+  local group = args[2]
+  local prefix = string.sub(idKey, 1, #idKey - 2)
+  local removed = 0
+
+  -- Build set of active entry IDs from PEL via paginated XPENDING
+  local activeSet = {}
+  local ok, pending = pcall(redis.call, 'XPENDING', streamKey, group, '-', '+', '10000')
+  if ok and pending and #pending > 0 then
+    for i = 1, #pending do
+      activeSet[pending[i][1]] = true
+    end
+    -- Page through remaining PEL entries if there were exactly 10000
+    while #pending == 10000 do
+      local lastId = pending[#pending][1]
+      local dashPos = lastId:find('-')
+      local seq = tonumber(lastId:sub(dashPos + 1))
+      local nextStart = lastId:sub(1, dashPos) .. tostring(seq + 1)
+      ok, pending = pcall(redis.call, 'XPENDING', streamKey, group, nextStart, '+', '10000')
+      if ok and pending and #pending > 0 then
+        for i = 1, #pending do
+          activeSet[pending[i][1]] = true
+        end
+      else
+        break
+      end
+    end
+  end
+
+  -- Paginated XRANGE to avoid loading entire stream into memory
+  local cursor = '-'
+  while true do
+    local entries = redis.call('XRANGE', streamKey, cursor, '+', 'COUNT', 1000)
+    if #entries == 0 then break end
+
+    local toDelete = {}
+    for i = 1, #entries do
+      local entryId = entries[i][1]
+      if not activeSet[entryId] then
+        toDelete[#toDelete + 1] = entryId
+        local fields = entries[i][2]
+        for j = 1, #fields, 2 do
+          if fields[j] == 'jobId' and fields[j + 1] ~= '' then
+            local jobId = fields[j + 1]
+            redis.call('UNLINK', prefix .. 'job:' .. jobId, prefix .. 'log:' .. jobId, prefix .. 'deps:' .. jobId, prefix .. 'jstream:' .. jobId, prefix .. 'signals:' .. jobId)
+            removed = removed + 1
+            break
+          end
+        end
+      end
+    end
+    if #toDelete > 0 then
+      for i = 1, #toDelete, 1000 do
+        redis.call('XDEL', streamKey, unpack(toDelete, i, math.min(i + 999, #toDelete)))
+      end
+    end
+
+    -- Advance cursor past the last entry
+    local lastId = entries[#entries][1]
+    local dashPos = lastId:find('-')
+    local seq = tonumber(lastId:sub(dashPos + 1))
+    cursor = lastId:sub(1, dashPos) .. tostring(seq + 1)
+  end
+
+  -- Optionally drain delayed/scheduled jobs
+  if drainDelayed then
+    local offset = 0
+    while true do
+      local scheduled = redis.call('ZRANGE', scheduledKey, offset, offset + 999)
+      if #scheduled == 0 then break end
+      local batch = {}
+      for j = 1, #scheduled do
+        local jobId = scheduled[j]
+        batch[#batch + 1] = prefix .. 'job:' .. jobId
+        batch[#batch + 1] = prefix .. 'log:' .. jobId
+        batch[#batch + 1] = prefix .. 'deps:' .. jobId
+        batch[#batch + 1] = prefix .. 'jstream:' .. jobId
+        batch[#batch + 1] = prefix .. 'signals:' .. jobId
+      end
+      redis.call('UNLINK', unpack(batch))
+      removed = removed + #scheduled
+      offset = offset + 1000
+    end
+    redis.call('UNLINK', scheduledKey)
+  end
+
+  -- Drain LIFO list: get all waiting job IDs, delete their hashes, then delete the list
+  if lifoKey and lifoKey ~= '' then
+    local lifoIds = redis.call('LRANGE', lifoKey, 0, -1)
+    for i = 1, #lifoIds do
+      local jobId = lifoIds[i]
+      redis.call('UNLINK', prefix .. 'job:' .. jobId, prefix .. 'log:' .. jobId, prefix .. 'deps:' .. jobId, prefix .. 'jstream:' .. jobId, prefix .. 'signals:' .. jobId)
+      removed = removed + 1
+    end
+    redis.call('UNLINK', lifoKey)
+  end
+
+  -- Drain priority list: get all waiting job IDs, delete their hashes, then delete the list
+  if priorityKey and priorityKey ~= '' then
+    local priorityIds = redis.call('LRANGE', priorityKey, 0, -1)
+    for i = 1, #priorityIds do
+      local jobId = priorityIds[i]
+      redis.call('UNLINK', prefix .. 'job:' .. jobId, prefix .. 'log:' .. jobId, prefix .. 'deps:' .. jobId, prefix .. 'jstream:' .. jobId, prefix .. 'signals:' .. jobId)
+      removed = removed + 1
+    end
+    redis.call('UNLINK', priorityKey)
+  end
+
+  if removed > 0 then
+    emitEvent(eventsKey, 'drained', tostring(removed), nil)
+  end
+  return removed
+end)
+
+redis.register_function('glidemq_retryJobs', function(keys, args)
+  local failedKey = keys[1]
+  local scheduledKey = keys[2]
+  local eventsKey = keys[3]
+  local idKey = keys[4]
+  local count = tonumber(args[1]) or 0
+  local timestamp = tonumber(args[2])
+  if not timestamp then return redis.error_reply('ERR invalid timestamp') end
+  local prefix = string.sub(idKey, 1, #idKey - 2)
+  local retried = 0
+
+  while true do
+    if count > 0 and retried >= count then break end
+    local batchSize = 1000
+    if count > 0 then
+      batchSize = math.min(1000, count - retried)
+    end
+    local ids = redis.call('ZRANGE', failedKey, 0, batchSize - 1)
+    if #ids == 0 then break end
+    redis.call('ZREM', failedKey, unpack(ids))
+    for i = 1, #ids do
+      local jobId = ids[i]
+      local jobKey = prefix .. 'job:' .. jobId
+      if redis.call('EXISTS', jobKey) == 1 then
+        local priority = tonumber(redis.call('HGET', jobKey, 'priority')) or 0
+        local score = priority * PRIORITY_SHIFT + timestamp
+        redis.call('ZADD', scheduledKey, score, jobId)
+        redis.call('HSET', jobKey,
+          'state', 'delayed',
+          'attemptsMade', '0',
+          'failedReason', '',
+          'finishedOn', ''
+        )
+        retried = retried + 1
+      end
+    end
+  end
+  if retried > 0 then
+    emitEvent(eventsKey, 'retried', tostring(retried), nil)
+  end
+  return retried
+end)
+
+redis.register_function('glidemq_healListActive', function(keys, args)
+  local idKey = keys[1]
+  local prefix = string.sub(idKey, 1, #idKey - 2)
+  local listActiveKey = prefix .. 'list-active'
+  local counter = tonumber(redis.call('GET', listActiveKey)) or 0
+  if counter <= 0 then
+    return 0
+  end
+  local pattern = prefix .. 'job:*'
+  local cursor = '0'
+  local actual = 0
+  local maxIter = 500
+  local iter = 0
+  repeat
+    iter = iter + 1
+    local scanResult = redis.call('SCAN', cursor, 'MATCH', pattern, 'COUNT', 100)
+    cursor = scanResult[1]
+    local scannedKeys = scanResult[2]
+    for i = 1, #scannedKeys do
+      local jk = scannedKeys[i]
+      local state = redis.call('HGET', jk, 'state')
+      if state == 'active' then
+        local lifo = redis.call('HGET', jk, 'lifo')
+        if lifo == '1' then
+          actual = actual + 1
+        else
+          local pri = tonumber(redis.call('HGET', jk, 'priority')) or 0
+          if pri > 0 then
+            actual = actual + 1
+          end
+        end
+      end
+    end
+  until cursor == '0' or iter >= maxIter
+  local drift = counter - actual
+  if drift > 0 then
+    redis.call('DECRBY', listActiveKey, drift)
+  end
+  return drift
+end)
+
+-- List active list-sourced jobIds (state='active' AND (lifo='1' OR priority>0)) via bounded SCAN.
+-- Stream-backed active jobs are in the consumer group PEL (XPENDING) and listed separately.
+-- All prefix:job:* keys share the {queueName} hash tag, so SCAN runs on a single
+-- cluster slot and observes the full set on its local node.
+-- KEYS: [idKey]
+-- ARGS: [start, end] (inclusive 0-indexed bounds; end < 0 means unbounded)
+-- Returns: array of jobId strings.
+redis.register_function('glidemq_getActiveListJobIds', function(keys, args)
+  local idKey = keys[1]
+  local prefix = string.sub(idKey, 1, #idKey - 2)
+  local startIdx = tonumber(args[1]) or 0
+  local endIdx = tonumber(args[2])
+  if endIdx == nil then endIdx = -1 end
+  local pattern = prefix .. 'job:*'
+  local cursor = '0'
+  local maxIter = 1000
+  local iter = 0
+  local ids = {}
+  repeat
+    iter = iter + 1
+    local sr = redis.call('SCAN', cursor, 'MATCH', pattern, 'COUNT', 500)
+    cursor = sr[1]
+    local scanned = sr[2]
+    for i = 1, #scanned do
+      local jk = scanned[i]
+      local state = redis.call('HGET', jk, 'state')
+      if state == 'active' then
+        local vals = redis.call('HMGET', jk, 'lifo', 'priority')
+        if vals[1] == '1' or (tonumber(vals[2]) or 0) > 0 then
+          ids[#ids + 1] = string.sub(jk, #prefix + 5)
+        end
+      end
+    end
+  until cursor == '0' or iter >= maxIter
+  if endIdx >= 0 then
+    local lastIdx = endIdx + 1
+    if lastIdx > #ids then lastIdx = #ids end
+    if startIdx + 1 > lastIdx then return {} end
+    local sliced = {}
+    for i = startIdx + 1, lastIdx do
+      sliced[#sliced + 1] = ids[i]
+    end
+    return sliced
+  end
+  if startIdx > 0 then
+    local sliced = {}
+    for i = startIdx + 1, #ids do
+      sliced[#sliced + 1] = ids[i]
+    end
+    return sliced
+  end
+  return ids
+end)
+
+redis.register_function('glidemq_popLists', function(keys, args)
+  local priorityKey = keys[1]
+  local lifoKey = keys[2]
+  local count = tonumber(args[1]) or 1
+  local results = {}
+  for i = 1, count do
+    local id = redis.call('RPOP', priorityKey)
+    if not id then break end
+    results[#results + 1] = id
+  end
+  if #results > 0 then return results end
+  for i = 1, count do
+    local id = redis.call('RPOP', lifoKey)
+    if not id then break end
+    results[#results + 1] = id
+  end
+  return results
+end)
+
+redis.register_function('glidemq_suspend', function(keys, args)
+  local jobKey = keys[1]
+  local streamKey = keys[2]
+  local eventsKey = keys[3]
+  local suspendedKey = keys[4]
+  local jobId = args[1]
+  local entryId = args[2]
+  local group = args[3]
+  local now = tonumber(args[4]) or 0
+  local reason = args[5] or ''
+  local timeout = tonumber(args[6]) or 0
+  local broadcastMode = args[7] or '0'
+
+  local state = redis.call('HGET', jobKey, 'state')
+  if not state then return 'error:not_found' end
+  if state ~= 'active' then return 'error:not_active' end
+
+  pcall(redis.call, 'XACK', streamKey, group, entryId)
+  if entryId ~= '' and broadcastMode ~= '1' then redis.call('XDEL', streamKey, entryId) end
+
+  local fields = {'state', 'suspended', 'suspendedAt', tostring(now)}
+  if reason ~= '' then
+    fields[#fields + 1] = 'suspendReason'
+    fields[#fields + 1] = reason
+  end
+  if timeout > 0 then
+    fields[#fields + 1] = 'suspendTimeout'
+    fields[#fields + 1] = tostring(timeout)
+  end
+  redis.call('HSET', jobKey, unpack(fields))
+
+  local deadline = timeout > 0 and (now + timeout) or 9999999999999
+  redis.call('ZADD', suspendedKey, deadline, jobId)
+
+  local ordSeq = tonumber(redis.call('HGET', jobKey, 'orderingSeq')) or 0
+  if ordSeq <= 0 then
+    releaseGroupSlotAndPromote(jobKey, jobId, now)
+  end
+
+  if entryId == '' then
+    local prefix = string.sub(jobKey, 1, #jobKey - #('job:' .. jobId))
+    decrListActive(prefix .. 'list-active')
+  end
+
+  emitEvent(eventsKey, 'suspended', jobId, nil)
+  return 'ok'
+end)
+
+redis.register_function('glidemq_signal', function(keys, args)
+  local jobKey = keys[1]
+  local streamKey = keys[2]
+  local eventsKey = keys[3]
+  local suspendedKey = keys[4]
+  local signalsKey = keys[5]
+  local jobId = args[1]
+  local signalName = args[2]
+  local signalData = args[3] or ''
+  local now = tonumber(args[4]) or 0
+
+  local state = redis.call('HGET', jobKey, 'state')
+  if not state or state ~= 'suspended' then return 'not_suspended' end
+
+  local signalJson = cjson.encode({name = signalName, data = signalData, receivedAt = now})
+  redis.call('RPUSH', signalsKey, signalJson)
+
+  local rawSignals = redis.call('LRANGE', signalsKey, 0, -1)
+  local signalsArr = {}
+  for i, s in ipairs(rawSignals) do signalsArr[i] = cjson.decode(s) end
+  redis.call('HSET', jobKey, 'signals', cjson.encode(signalsArr))
+
+  redis.call('HSET', jobKey, 'state', 'waiting')
+  redis.call('ZREM', suspendedKey, jobId)
+
+  local jobName = redis.call('HGET', jobKey, 'name') or ''
+  xaddJob(streamKey, jobId, jobName)
+
+  emitEvent(eventsKey, 'resumed', jobId, {'signal', signalName})
+  return 'ok'
+end)
+
+redis.register_function('glidemq_sweepSuspended', function(keys, args)
+  local suspendedKey = keys[1]
+  local eventsKey = keys[2]
+  local failedKey = keys[3]
+  local metricsKey = keys[4]
+  local now = tonumber(args[1]) or 0
+  local keyPrefix = args[2] or ''
+
+  local expired = redis.call('ZRANGEBYSCORE', suspendedKey, '0', string.format('%.0f', now), 'LIMIT', 0, 100)
+  if not expired or #expired == 0 then return 0 end
+
+  local count = 0
+  for _, jobId in ipairs(expired) do
+    local id = tostring(jobId)
+    local jobKey = keyPrefix .. 'job:' .. id
+    local jState = redis.call('HGET', jobKey, 'state')
+    if jState == 'suspended' then
+      local processedOn = tonumber(redis.call('HGET', jobKey, 'processedOn')) or now
+      local ordKey = redis.call('HGET', jobKey, 'orderingKey')
+      local ordSeq = tonumber(redis.call('HGET', jobKey, 'orderingSeq')) or 0
+      local grpKey = redis.call('HGET', jobKey, 'groupKey')
+      redis.call('ZADD', failedKey, now, id)
+      redis.call('HSET', jobKey,
+        'state', 'failed',
+        'failedReason', 'Suspend timeout exceeded',
+        'finishedOn', tostring(now),
+        'processedOn', tostring(now)
+      )
+      markOrderingDone(jobKey, id, ordKey, ordSeq)
+      -- Only release the group slot for ordered jobs. Non-ordered jobs already
+      -- released their slot in glidemq_suspend; releasing again would double-
+      -- decrement the group's active counter.
+      if ordSeq > 0 then
+        releaseGroupSlotAndPromote(jobKey, id, now, grpKey)
+      end
+      emitEvent(eventsKey, 'failed', id, {'failedReason', 'Suspend timeout exceeded'})
+      recordMetrics(metricsKey, now, now - processedOn)
+      count = count + 1
+    end
+    redis.call('ZREM', suspendedKey, id)
+  end
+  return count
+end)
+
+redis.register_function('glidemq_checkBudget', function(keys, args)
+  local budgetKey = keys[1]
+  local exists = redis.call('EXISTS', budgetKey)
+  if exists == 0 then return 'no_budget' end
+  local exceeded = redis.call('HGET', budgetKey, 'exceeded')
+  if exceeded == '1' then return 'exceeded' end
+  return 'ok'
+end)
+
+redis.register_function('glidemq_recordUsageAndCheckBudget', function(keys, args)
+  local budgetKey = keys[1]
+  local tokensJson = args[1]
+  local costsJson = args[2]
+  local weightedTotal = tonumber(args[3]) or 0
+  local totalCost = tonumber(args[4]) or 0
+  local maxTokensJson = args[5]
+  local maxCostsJson = args[6]
+
+  local exists = redis.call('EXISTS', budgetKey)
+  if exists == 0 then return 'no_budget' end
+
+  -- Increment weighted total and total cost
+  if weightedTotal > 0 then redis.call('HINCRBYFLOAT', budgetKey, 'usedTokens', weightedTotal) end
+  if totalCost > 0 then redis.call('HINCRBYFLOAT', budgetKey, 'usedCost', totalCost) end
+
+  -- Increment per-category token counters
+  local ok1, tokens = pcall(cjson.decode, tokensJson)
+  if ok1 and type(tokens) == 'table' then
+    for cat, val in pairs(tokens) do
+      if tonumber(val) and tonumber(val) > 0 then
+        redis.call('HINCRBYFLOAT', budgetKey, 'usedTokens:' .. cat, val)
+      end
+    end
+  end
+
+  -- Increment per-category cost counters
+  local ok2, costs = pcall(cjson.decode, costsJson)
+  if ok2 and type(costs) == 'table' then
+    for cat, val in pairs(costs) do
+      if tonumber(val) and tonumber(val) > 0 then
+        redis.call('HINCRBYFLOAT', budgetKey, 'usedCost:' .. cat, val)
+      end
+    end
+  end
+
+  -- Check weighted total tokens cap
+  local maxTotalTokens = tonumber(redis.call('HGET', budgetKey, 'maxTotalTokens')) or 0
+  local usedTokens = tonumber(redis.call('HGET', budgetKey, 'usedTokens')) or 0
+  if maxTotalTokens > 0 and usedTokens > maxTotalTokens then
+    redis.call('HSET', budgetKey, 'exceeded', '1')
+    return 'exceeded'
+  end
+
+  -- Check total cost cap
+  local maxTotalCost = tonumber(redis.call('HGET', budgetKey, 'maxTotalCost')) or 0
+  local usedCost = tonumber(redis.call('HGET', budgetKey, 'usedCost')) or 0
+  if maxTotalCost > 0 and usedCost > maxTotalCost then
+    redis.call('HSET', budgetKey, 'exceeded', '1')
+    return 'exceeded'
+  end
+
+  -- Check per-category token limits
+  local ok3, maxTokens = pcall(cjson.decode, maxTokensJson)
+  if ok3 and type(maxTokens) == 'table' then
+    for cat, limit in pairs(maxTokens) do
+      local used = tonumber(redis.call('HGET', budgetKey, 'usedTokens:' .. cat)) or 0
+      if tonumber(limit) and tonumber(limit) > 0 and used > tonumber(limit) then
+        redis.call('HSET', budgetKey, 'exceeded', '1')
+        return 'exceeded'
+      end
+    end
+  end
+
+  -- Check per-category cost limits
+  local ok4, maxCosts = pcall(cjson.decode, maxCostsJson)
+  if ok4 and type(maxCosts) == 'table' then
+    for cat, limit in pairs(maxCosts) do
+      local used = tonumber(redis.call('HGET', budgetKey, 'usedCost:' .. cat)) or 0
+      if tonumber(limit) and tonumber(limit) > 0 and used > tonumber(limit) then
+        redis.call('HSET', budgetKey, 'exceeded', '1')
+        return 'exceeded'
+      end
+    end
+  end
+
+  return 'ok'
 end)
